@@ -82,24 +82,24 @@
         "deploy:ci": "ng deploy --no-build --message=\"Release $npm_package_name (v$npm_package_version) on gh-pages\""
     },
     "dependencies": {
-        "@angular/animations": "^15.2.9",
-        "@angular/common": "^15.2.9",
-        "@angular/compiler": "^15.2.9",
-        "@angular/core": "^15.2.9",
-        "@angular/forms": "^15.2.9",
-        "@angular/localize": "^15.2.9",
-        "@angular/platform-browser": "^15.2.9",
-        "@angular/platform-browser-dynamic": "^15.2.9",
-        "@angular/platform-server": "^15.2.9",
-        "@angular/router": "^15.2.9",
-        "@codemirror/legacy-modes": "^6.3.2",
-        "@fortawesome/angular-fontawesome": "^0.12.1",
-        "@fortawesome/fontawesome-svg-core": "^6.4.0",
-        "@fortawesome/free-solid-svg-icons": "^6.4.0",
+        "@angular/animations": "^17.0.7",
+        "@angular/common": "^17.0.7",
+        "@angular/compiler": "^17.0.7",
+        "@angular/core": "^17.0.7",
+        "@angular/forms": "^17.0.7",
+        "@angular/localize": "^17.0.7",
+        "@angular/platform-browser": "^17.0.7",
+        "@angular/platform-browser-dynamic": "^17.0.7",
+        "@angular/platform-server": "^17.0.7",
+        "@angular/router": "^17.0.7",
+        "@codemirror/legacy-modes": "^6.3.3",
+        "@fortawesome/angular-fontawesome": "^0.14.0",
+        "@fortawesome/fontawesome-svg-core": "^6.5.1",
+        "@fortawesome/free-solid-svg-icons": "^6.5.1",
         "@kolkov/ngx-gallery": "2.0.1",
-        "@ng-bootstrap/ng-bootstrap": "^14.1.1",
-        "@popperjs/core": "^2.11.7",
-        "bootstrap": "^5.2.3",
+        "@ng-bootstrap/ng-bootstrap": "^16.0.0",
+        "@popperjs/core": "^2.11.8",
+        "bootstrap": "^5.3.2",
         "codemirror": "^6.0.1",
         "d3-drag": "^3.0.0",
         "d3-fetch": "^3.0.1",
@@ -108,58 +108,59 @@
         "d3-zoom": "^3.0.0",
         "font-awesome": "^4.7.0",
         "json2typescript": "^1.5.1",
-        "n3": "^1.16.4",
+        "n3": "^1.17.2",
         "ngx-json-viewer": "^3.2.1",
         "rdfstore": "musicenfanthen/rdfstore-js#v0.9.18-alpha.12",
         "rxjs": "~7.8.1",
         "snapsvg": "^0.5.1",
-        "tslib": "^2.5.0",
-        "zone.js": "~0.13.0"
+        "stream": "^0.0.2",
+        "tslib": "^2.6.2",
+        "zone.js": "~0.14.2"
     },
     "devDependencies": {
-        "@angular-devkit/build-angular": "^15.2.8",
-        "@angular-eslint/builder": "^15.2.1",
-        "@angular-eslint/eslint-plugin": "^15.2.1",
-        "@angular-eslint/eslint-plugin-template": "^15.2.1",
-        "@angular-eslint/schematics": "^15.2.1",
-        "@angular-eslint/template-parser": "^15.2.1",
-        "@angular/cli": "^15.2.8",
-        "@angular/compiler-cli": "^15.2.9",
-        "@commitlint/cli": "^17.6.3",
-        "@commitlint/config-angular": "^17.6.3",
-        "@compodoc/compodoc": "^1.1.19",
-        "@types/d3": "^7.4.0",
-        "@types/jasmine": "~4.3.1",
-        "@types/node": "^16.18.27",
-        "@typescript-eslint/eslint-plugin": "^5.59.5",
-        "@typescript-eslint/parser": "^5.59.5",
-        "angular-cli-ghpages": "^1.0.5",
-        "conventional-recommended-bump": "^6.1.0",
-        "eslint": "^8.40.0",
-        "eslint-config-prettier": "^8.8.0",
+        "@angular-devkit/build-angular": "^17.0.7",
+        "@angular-eslint/builder": "^17.1.1",
+        "@angular-eslint/eslint-plugin": "^17.1.1",
+        "@angular-eslint/eslint-plugin-template": "^17.1.1",
+        "@angular-eslint/schematics": "^17.1.1",
+        "@angular-eslint/template-parser": "^17.1.1",
+        "@angular/cli": "^17.0.7",
+        "@angular/compiler-cli": "^17.0.7",
+        "@commitlint/cli": "^18.4.3",
+        "@commitlint/config-angular": "^18.4.3",
+        "@compodoc/compodoc": "^1.1.23",
+        "@types/d3": "^7.4.3",
+        "@types/jasmine": "~5.1.4",
+        "@types/node": "^18.19.3",
+        "@typescript-eslint/eslint-plugin": "^6.14.0",
+        "@typescript-eslint/parser": "^6.14.0",
+        "angular-cli-ghpages": "^1.0.7",
+        "conventional-recommended-bump": "^9.0.0",
+        "eslint": "^8.56.0",
+        "eslint-config-prettier": "^9.1.0",
         "eslint-plugin-angular": "^4.1.0",
-        "eslint-plugin-deprecation": "^1.4.1",
-        "eslint-plugin-import": "^2.27.5",
-        "eslint-plugin-jsdoc": "^43.0.7",
-        "eslint-plugin-prettier": "^4.2.1",
+        "eslint-plugin-deprecation": "^2.0.0",
+        "eslint-plugin-import": "^2.29.1",
+        "eslint-plugin-jsdoc": "^46.9.1",
+        "eslint-plugin-prettier": "^5.0.1",
         "gzipper": "^7.2.0",
         "husky": "^8.0.3",
-        "jasmine-core": "~4.6.0",
+        "jasmine-core": "~5.1.1",
         "karma": "~6.4.2",
         "karma-chrome-launcher": "~3.2.0",
-        "karma-coverage": "^2.2.0",
+        "karma-coverage": "^2.2.1",
         "karma-jasmine": "~5.1.0",
-        "karma-jasmine-html-reporter": "^2.0.0",
-        "lint-staged": "^13.2.2",
-        "prettier": "^2.8.8",
+        "karma-jasmine-html-reporter": "^2.1.0",
+        "lint-staged": "^15.2.0",
+        "prettier": "^3.1.1",
         "source-map-explorer": "^2.5.3",
         "standard-version": "^9.5.0",
-        "typescript": "~4.9.5",
-        "webpack-bundle-analyzer": "^4.8.0"
+        "typescript": "~5.2.2",
+        "webpack-bundle-analyzer": "^4.10.1"
     },
     "engines": {
-        "node": ">= 14.17.0",
-        "npm": ">= 6.14.13",
+        "node": ">= 18.13.0",
+        "npm": ">= 8.19.3",
         "yarn": "^1.22.0"
     }
 }
