@@ -39,7 +39,11 @@ export class EditionComplexComponent implements OnDestroy, OnInit {
      * @param {EditionService} editionService Instance of the EditionService.
      * @param {UtilityService} utils Instance of the UtilityService.
      */
-    constructor(private route: ActivatedRoute, private editionService: EditionService, public utils: UtilityService) {}
+    constructor(
+        private route: ActivatedRoute,
+        private editionService: EditionService,
+        public utils: UtilityService
+    ) {}
 
     /**
      * Getter variable: editionRouteConstants.
