@@ -4,13 +4,16 @@ import { ComponentFixture, TestBed, waitForAsync } from '@angular/core/testing';
 import { EMPTY, Observable, of as observableOf } from 'rxjs';
 import Spy = jasmine.Spy;
 
-import { NgbAccordion, NgbAccordionModule, NgbConfig, NgbPanelChangeEvent } from '@ng-bootstrap/ng-bootstrap';
+import { NgbAccordionDirective, NgbAccordionModule, NgbConfig } from '@ng-bootstrap/ng-bootstrap';
 
 import { click } from '@testing/click-helper';
 import { detectChangesOnPush } from '@testing/detect-changes-on-push-helper';
 
 import {
     expectSpyCall,
+    expectToBe,
+    expectToContain,
+    expectToEqual,
     getAndExpectDebugElementByCss,
     getAndExpectDebugElementByDirective,
 } from '@testing/expect-helper';
@@ -38,13 +41,14 @@ describe('ConstructResultsComponent (DONE)', () => {
     let compDe: DebugElement;
 
     let expectedHeight: number;
-    let expectedTriples: Triple[];
+    let expectedQueryResult: Triple;
     let expectedQueryResult$: Observable<Triple[]>;
     let expectedIsFullscreen: boolean;
 
     let emitClickedNodeRequestSpy: Spy;
+    let isAccordionItemDisabledSpy: Spy;
+    let isQueryResultNotEmptySpy: Spy;
     let nodeClickSpy: Spy;
-    let preventPanelCollapseOnFullscreenSpy: Spy;
 
     // Global NgbConfigModule
     @NgModule({ imports: [NgbAccordionModule], exports: [NgbAccordionModule] })
@@ -57,7 +61,7 @@ describe('ConstructResultsComponent (DONE)', () => {
 
     beforeEach(waitForAsync(() => {
         TestBed.configureTestingModule({
-            imports: [NgbAccordionWithConfigModule],
+            imports: [NgbAccordionWithConfigModule, NgbAccordionDirective],
             declarations: [
                 ConstructResultsComponent,
                 ForceGraphStubComponent,
@@ -75,21 +79,20 @@ describe('ConstructResultsComponent (DONE)', () => {
         // Test data
         expectedHeight = 500;
         expectedIsFullscreen = false;
-        expectedTriples = [
-            {
-                subject: 'example:Test',
-                predicate: 'example:has',
-                object: 'example:Success',
-            },
-        ];
-        expectedQueryResult$ = observableOf(expectedTriples);
+        expectedQueryResult = {
+            subject: 'example:Test',
+            predicate: 'example:has',
+            object: 'example:Success',
+        };
+        expectedQueryResult$ = observableOf([expectedQueryResult]);
 
         // Spies on component functions
         // `.and.callThrough` will track the spy down the nested describes, see
         // https://jasmine.github.io/2.0/introduction.html#section-Spies:_%3Ccode%3Eand.callThrough%3C/code%3E
-        nodeClickSpy = spyOn(component, 'onGraphNodeClick').and.callThrough();
-        preventPanelCollapseOnFullscreenSpy = spyOn(component, 'preventPanelCollapseOnFullscreen').and.callThrough();
         emitClickedNodeRequestSpy = spyOn(component.clickedNodeRequest, 'emit').and.callThrough();
+        isAccordionItemDisabledSpy = spyOn(component, 'isAccordionItemDisabled').and.callThrough();
+        isQueryResultNotEmptySpy = spyOn(component, 'isQueryResultNotEmpty').and.callThrough();
+        nodeClickSpy = spyOn(component, 'onGraphNodeClick').and.callThrough();
     });
 
     it('... should create', () => {
@@ -110,12 +113,9 @@ describe('ConstructResultsComponent (DONE)', () => {
         });
 
         describe('VIEW', () => {
-            it('... should contain one ngb-accordion without panel (div.accordion-item) yet', () => {
-                // Ngb-accordion debug element
-                const accordionDes = getAndExpectDebugElementByCss(compDe, 'ngb-accordion', 1, 1);
-
-                // Panel
-                getAndExpectDebugElementByCss(accordionDes[0], 'div.accordion-item', 0, 0, 'yet');
+            it('... should contain no div.accordion yet', () => {
+                // Div.accordion debug element
+                getAndExpectDebugElementByCss(compDe, 'div.accordion', 0, 0);
             });
         });
     });
@@ -132,211 +132,909 @@ describe('ConstructResultsComponent (DONE)', () => {
         });
 
         it('... should have `queryResult` input', () => {
-            expect(component.queryResult$).toBeDefined();
-            expect(component.queryResult$)
-                .withContext(`should equal ${expectedQueryResult$}`)
-                .toEqual(expectedQueryResult$);
+            expectToEqual(component.queryResult$, expectedQueryResult$);
         });
 
         it('... should have `defaultForceGraphHeight` input', () => {
-            expect(component.defaultForceGraphHeight).toBeDefined();
-            expect(component.defaultForceGraphHeight).withContext(`should be ${expectedHeight}`).toBe(expectedHeight);
+            expectToBe(component.defaultForceGraphHeight, expectedHeight);
         });
 
         it('... should have `isFullscreen` input', () => {
-            expect(component.isFullscreen).toBeDefined();
-            expect(component.isFullscreen).withContext(`should be ${expectedIsFullscreen}`).toBe(expectedIsFullscreen);
+            expectToBe(component.isFullscreen, expectedIsFullscreen);
         });
 
         describe('VIEW', () => {
-            it('... should contain one ngb-accordion with panel (div.accordion-item) header and body', () => {
-                // Ngb-accordion debug element
-                const accordionDes = getAndExpectDebugElementByCss(compDe, 'ngb-accordion', 1, 1);
-
-                // Panel (div.card)
-                const panelDes = getAndExpectDebugElementByCss(accordionDes[0], 'div.accordion-item', 1, 1); // Panel (div.card)
-                // Header
-                getAndExpectDebugElementByCss(
-                    panelDes[0],
-                    'div#awg-graph-visualizer-construct-results-header.accordion-header',
-                    1,
-                    1
-                ); // Panel (div.card)
-                // Body
-                getAndExpectDebugElementByCss(
-                    panelDes[0],
-                    'div#awg-graph-visualizer-construct-results > div.accordion-body',
-                    1,
-                    1
-                );
+            it('... should contain one div.accordion', () => {
+                // NgbAccordion debug element
+                getAndExpectDebugElementByCss(compDe, 'div.accordion', 1, 1);
             });
 
-            it('... should display panel header button', () => {
-                // Panel header button
-                const btnDes = getAndExpectDebugElementByCss(
-                    compDe,
-                    'div#awg-graph-visualizer-construct-results-header > button',
-                    1,
-                    1
-                );
+            describe('not in fullscreen mode', () => {
+                it('... should contain div.accordion-item with header and open body in div.accordion', () => {
+                    // NgbAccordion debug element
+                    const accordionDes = getAndExpectDebugElementByCss(compDe, 'div.accordion', 1, 1);
 
-                const btnEl = btnDes[0].nativeElement;
+                    // Div.accordion-item
+                    const itemDes = getAndExpectDebugElementByCss(
+                        accordionDes[0],
+                        'div#awg-graph-visualizer-construct-results.accordion-item',
+                        1,
+                        1
+                    );
+                    // Header (div.accordion-header)
+                    getAndExpectDebugElementByCss(
+                        itemDes[0],
+                        'div#awg-graph-visualizer-construct-results > div.accordion-header',
+                        1,
+                        1
+                    );
 
-                // Check button content
-                expect(btnEl.textContent).toBeTruthy();
-                expect(btnEl.textContent).withContext('should be Resultat').toContain('Resultat');
+                    // Body open (div.accordion-collapse)
+                    const itemBodyDes = getAndExpectDebugElementByCss(
+                        itemDes[0],
+                        'div#awg-graph-visualizer-construct-results-collapse',
+                        1,
+                        1
+                    );
+                    const itemBodyEl = itemBodyDes[0].nativeElement;
+
+                    expectToContain(itemBodyEl.classList, 'show');
+                });
+
+                it('... should display item header button', () => {
+                    // Header debug elements
+                    const itemHeaderDes = getAndExpectDebugElementByCss(
+                        compDe,
+                        'div#awg-graph-visualizer-construct-results > div.accordion-header',
+                        1,
+                        1
+                    );
+
+                    // Item header button
+                    const btnDes = getAndExpectDebugElementByCss(itemHeaderDes[0], 'button.accordion-button', 1, 1);
+                    const btnEl = btnDes[0].nativeElement;
+
+                    // Check button content
+                    expectToBe(btnEl.textContent, 'Resultat');
+                });
+
+                it('... should toggle item body on click', () => {
+                    // Header debug elements
+                    const itemHeaderDes = getAndExpectDebugElementByCss(
+                        compDe,
+                        'div#awg-graph-visualizer-construct-results > div.accordion-header',
+                        1,
+                        1
+                    );
+
+                    // Button debug elements
+                    const btnDes = getAndExpectDebugElementByCss(itemHeaderDes[0], 'button.accordion-button', 1, 1);
+                    // Button native elements to click on
+                    const btnEl = btnDes[0].nativeElement;
+
+                    // Item body is open
+                    let itemBodyDes = getAndExpectDebugElementByCss(
+                        compDe,
+                        'div#awg-graph-visualizer-construct-results > div.accordion-collapse',
+                        1,
+                        1,
+                        'open'
+                    );
+                    let itemBodyEl = itemBodyDes[0].nativeElement;
+
+                    expectToContain(itemBodyEl.classList, 'show');
+
+                    // Click header button
+                    click(btnEl as HTMLElement);
+                    detectChangesOnPush(fixture);
+
+                    // Item body is collapsed
+                    itemBodyDes = getAndExpectDebugElementByCss(
+                        compDe,
+                        'div#awg-graph-visualizer-construct-results > div.accordion-collapse',
+                        1,
+                        1,
+                        'collapsed'
+                    );
+                    itemBodyEl = itemBodyDes[0].nativeElement;
+
+                    expectToContain(itemBodyEl.classList, 'collapse');
+
+                    // Click header button
+                    click(btnEl as HTMLElement);
+                    detectChangesOnPush(fixture);
+
+                    // Item body is open again
+                    itemBodyDes = getAndExpectDebugElementByCss(
+                        compDe,
+                        'div#awg-graph-visualizer-construct-results > div.accordion-collapse',
+                        1,
+                        1,
+                        'open'
+                    );
+                    itemBodyEl = itemBodyDes[0].nativeElement;
+
+                    expectToContain(itemBodyEl.classList, 'show');
+                });
+
+                describe('... should contain TwelveToneSpinnerComponent (stubbed) in item body while loading if ... ', () => {
+                    it('... queryResult$ is EMPTY', () => {
+                        // Mock empty observable
+                        component.queryResult$ = EMPTY;
+                        detectChangesOnPush(fixture);
+
+                        // Item body
+                        const bodyDes = getAndExpectDebugElementByCss(
+                            compDe,
+                            'div#awg-graph-visualizer-construct-results-collapse > div.accordion-body',
+                            1,
+                            1
+                        );
+
+                        getAndExpectDebugElementByDirective(bodyDes[0], TwelveToneSpinnerStubComponent, 1, 1);
+                    });
+
+                    it('... queryResult$ is undefined', () => {
+                        // Mock undefined response
+                        component.queryResult$ = observableOf(undefined);
+                        detectChangesOnPush(fixture);
+
+                        // Item body
+                        const bodyDes = getAndExpectDebugElementByCss(
+                            compDe,
+                            'div#awg-graph-visualizer-construct-results-collapse > div.accordion-body',
+                            1,
+                            1
+                        );
+
+                        getAndExpectDebugElementByDirective(bodyDes[0], TwelveToneSpinnerStubComponent, 1, 1);
+                    });
+
+                    it('... queryResult$ is null', () => {
+                        // Mock null response
+                        component.queryResult$ = observableOf(null);
+                        detectChangesOnPush(fixture);
+
+                        // Item body
+                        const bodyDes = getAndExpectDebugElementByCss(
+                            compDe,
+                            'div#awg-graph-visualizer-construct-results-collapse > div.accordion-body',
+                            1,
+                            1
+                        );
+
+                        getAndExpectDebugElementByDirective(bodyDes[0], TwelveToneSpinnerStubComponent, 1, 1);
+                    });
+                });
+
+                describe('... should contain item body with SparqlNoResultsStubComponent (stubbed) if ... ', () => {
+                    it('... queryResult is empty array', () => {
+                        // Mock empty array
+                        component.queryResult$ = observableOf([]);
+                        detectChangesOnPush(fixture);
+
+                        // Item body
+                        const bodyDes = getAndExpectDebugElementByCss(
+                            compDe,
+                            'div#awg-graph-visualizer-construct-results-collapse > div.accordion-body',
+                            1,
+                            1
+                        );
+
+                        getAndExpectDebugElementByDirective(bodyDes[0], SparqlNoResultsStubComponent, 1, 1);
+                    });
+
+                    it('... queryResult is undefined or empty string', () => {
+                        // Mock undefined response
+                        component.queryResult$ = observableOf([
+                            { subject: undefined, predicate: undefined, object: undefined },
+                        ]);
+                        detectChangesOnPush(fixture);
+
+                        // Item body
+                        const bodyDes = getAndExpectDebugElementByCss(
+                            compDe,
+                            'div#awg-graph-visualizer-construct-results-collapse > div.accordion-body',
+                            1,
+                            1
+                        );
+
+                        getAndExpectDebugElementByDirective(bodyDes[0], SparqlNoResultsStubComponent, 1, 1);
+
+                        // Mock empty response
+                        component.queryResult$ = observableOf([{ subject: '', predicate: '', object: '' }]);
+                        detectChangesOnPush(fixture);
+
+                        // Item body
+                        getAndExpectDebugElementByCss(
+                            compDe,
+                            'div#awg-graph-visualizer-construct-results-collapse > div.accordion-body',
+                            1,
+                            1
+                        );
+
+                        getAndExpectDebugElementByDirective(bodyDes[0], SparqlNoResultsStubComponent, 1, 1);
+                    });
+
+                    it('... queryResult.subject is undefined or empty string', () => {
+                        // Mock undefined response
+                        component.queryResult$ = observableOf([
+                            {
+                                subject: undefined,
+                                predicate: 'example:has',
+                                object: 'example:Success',
+                            },
+                        ]);
+                        detectChangesOnPush(fixture);
+
+                        // Item body
+                        const bodyDes = getAndExpectDebugElementByCss(
+                            compDe,
+                            'div#awg-graph-visualizer-construct-results-collapse > div.accordion-body',
+                            1,
+                            1
+                        );
+
+                        getAndExpectDebugElementByDirective(bodyDes[0], SparqlNoResultsStubComponent, 1, 1);
+
+                        // Mock empty response
+                        component.queryResult$ = observableOf([
+                            {
+                                subject: '',
+                                predicate: 'example:has',
+                                object: 'example:Success',
+                            },
+                        ]);
+                        detectChangesOnPush(fixture);
+
+                        // Item body
+                        getAndExpectDebugElementByCss(
+                            compDe,
+                            'div#awg-graph-visualizer-construct-results-collapse > div.accordion-body',
+                            1,
+                            1
+                        );
+
+                        getAndExpectDebugElementByDirective(bodyDes[0], SparqlNoResultsStubComponent, 1, 1);
+                    });
+
+                    it('... queryResult.predicate is undefined or empty string', () => {
+                        // Mock undefined response
+                        component.queryResult$ = observableOf([
+                            {
+                                subject: 'example:Test',
+                                predicate: undefined,
+                                object: 'example:Success',
+                            },
+                        ]);
+                        detectChangesOnPush(fixture);
+
+                        // Item body
+                        const bodyDes = getAndExpectDebugElementByCss(
+                            compDe,
+                            'div#awg-graph-visualizer-construct-results-collapse > div.accordion-body',
+                            1,
+                            1
+                        );
+
+                        getAndExpectDebugElementByDirective(bodyDes[0], SparqlNoResultsStubComponent, 1, 1);
+
+                        // Mock empty response
+                        component.queryResult$ = observableOf([
+                            {
+                                subject: 'example:Test',
+                                predicate: '',
+                                object: 'example:Success',
+                            },
+                        ]);
+                        detectChangesOnPush(fixture);
+
+                        // Item body
+                        getAndExpectDebugElementByCss(
+                            compDe,
+                            'div#awg-graph-visualizer-construct-results-collapse > div.accordion-body',
+                            1,
+                            1
+                        );
+
+                        getAndExpectDebugElementByDirective(bodyDes[0], SparqlNoResultsStubComponent, 1, 1);
+                    });
+
+                    it('... queryResult.object is undefined or empty string', () => {
+                        // Mock undefined response
+                        component.queryResult$ = observableOf([
+                            {
+                                subject: 'example:Test',
+                                predicate: 'example:has',
+                                object: undefined,
+                            },
+                        ]);
+                        detectChangesOnPush(fixture);
+
+                        // Item body
+                        const bodyDes = getAndExpectDebugElementByCss(
+                            compDe,
+                            'div#awg-graph-visualizer-construct-results-collapse > div.accordion-body',
+                            1,
+                            1
+                        );
+
+                        getAndExpectDebugElementByDirective(bodyDes[0], SparqlNoResultsStubComponent, 1, 1);
+
+                        // Mock empty response
+                        component.queryResult$ = observableOf([
+                            {
+                                subject: 'example:Test',
+                                predicate: 'example:has',
+                                object: '',
+                            },
+                        ]);
+                        detectChangesOnPush(fixture);
+
+                        // Item body
+                        getAndExpectDebugElementByCss(
+                            compDe,
+                            'div#awg-graph-visualizer-construct-results-collapse > div.accordion-body',
+                            1,
+                            1
+                        );
+
+                        getAndExpectDebugElementByDirective(bodyDes[0], SparqlNoResultsStubComponent, 1, 1);
+                    });
+                });
+
+                it('... should contain item body with ForceGraphComponent (stubbed) if results are available', () => {
+                    // Item body
+                    const bodyDes = getAndExpectDebugElementByCss(
+                        compDe,
+                        'div#awg-graph-visualizer-construct-results-collapse > div.accordion-body',
+                        1,
+                        1
+                    );
+
+                    // ForceGraph
+                    getAndExpectDebugElementByDirective(bodyDes[0], ForceGraphStubComponent, 1, 1);
+                });
+
+                it('... should pass down `queryResult` and `defaultForceGraphHeight` to forceGraph component', () => {
+                    const forceGraphDes = getAndExpectDebugElementByDirective(compDe, ForceGraphStubComponent, 1, 1);
+                    const forceGraphCmp = forceGraphDes[0].injector.get(
+                        ForceGraphStubComponent
+                    ) as ForceGraphStubComponent;
+
+                    expectToEqual(forceGraphCmp.queryResultTriples, [expectedQueryResult]);
+                    expectToBe(forceGraphCmp.height, expectedHeight);
+                });
             });
 
-            it('... should toggle panel body on click when not in fullscreen mode', () => {
-                component.isFullscreen = false;
-                fixture.detectChanges();
+            describe('in fullscreen mode', () => {
+                beforeEach(() => {
+                    // Set fullscreen mode
+                    component.isFullscreen = true;
+                    detectChangesOnPush(fixture);
+                });
 
-                // Header debug elements
-                const panelHeaderDes = getAndExpectDebugElementByCss(
-                    compDe,
-                    'div#awg-graph-visualizer-construct-results-header.accordion-header',
-                    1,
-                    1
-                );
+                it('... should contain div.accordion-item with header and open body in div.accordion', () => {
+                    // NgbAccordion debug element
+                    const accordionDes = getAndExpectDebugElementByCss(compDe, 'div.accordion', 1, 1);
 
-                // Button debug elements
-                const btnDes = getAndExpectDebugElementByCss(panelHeaderDes[0], 'button.accordion-button', 1, 1);
-                // Button native elements to click on
-                const btnEl = btnDes[0].nativeElement;
+                    // Panel (div.accordion-item)
+                    const itemDes = getAndExpectDebugElementByCss(
+                        accordionDes[0],
+                        'div#awg-graph-visualizer-construct-results.accordion-item',
+                        1,
+                        1
+                    );
+                    // Header (div.accordion-header)
+                    getAndExpectDebugElementByCss(
+                        itemDes[0],
+                        'div#awg-graph-visualizer-construct-results > div.accordion-header',
+                        1,
+                        1
+                    );
 
-                // Panel body is open
-                getAndExpectDebugElementByCss(
-                    compDe,
-                    'div#awg-graph-visualizer-construct-results > div.accordion-body',
-                    1,
-                    1,
-                    'open'
-                );
+                    // Body open (div.accordion-collapse)
+                    const itemBodyDes = getAndExpectDebugElementByCss(
+                        itemDes[0],
+                        'div#awg-graph-visualizer-construct-results-collapse',
+                        1,
+                        1
+                    );
+                    const itemBodyEl = itemBodyDes[0].nativeElement;
 
-                // Click header button
-                click(btnEl as HTMLElement);
-                detectChangesOnPush(fixture);
+                    expectToContain(itemBodyEl.classList, 'show');
+                });
 
-                // Panel is open
-                getAndExpectDebugElementByCss(
-                    compDe,
-                    'div#awg-graph-visualizer-construct-results > div.accordion-body',
-                    0,
-                    0,
-                    'collapsed'
-                );
+                it('... should display item header button', () => {
+                    // Header debug elements
+                    const itemHeaderDes = getAndExpectDebugElementByCss(
+                        compDe,
+                        'div#awg-graph-visualizer-construct-results > div.accordion-header',
+                        1,
+                        1
+                    );
 
-                // Click header button
-                click(btnEl as HTMLElement);
-                detectChangesOnPush(fixture);
+                    // Panel header button
+                    const btnDes = getAndExpectDebugElementByCss(itemHeaderDes[0], 'button.accordion-button', 1, 1);
+                    const btnEl = btnDes[0].nativeElement;
 
-                getAndExpectDebugElementByCss(
-                    // Panel body is closed again
-                    compDe,
-                    'div#awg-graph-visualizer-construct-results > div.accordion-body',
-                    1,
-                    1,
-                    'open'
-                );
+                    // Check button content
+                    expectToBe(btnEl.textContent, 'Resultat');
+                });
+
+                it('... should not toggle item body on click', () => {
+                    // Header debug elements
+                    const itemHeaderDes = getAndExpectDebugElementByCss(
+                        compDe,
+                        'div#awg-graph-visualizer-construct-results > div.accordion-header',
+                        1,
+                        1
+                    );
+
+                    // Button debug elements
+                    const btnDes = getAndExpectDebugElementByCss(itemHeaderDes[0], 'button.accordion-button', 1, 1);
+                    // Button native elements to click on
+                    const btnEl = btnDes[0].nativeElement;
+
+                    expect(btnEl.disabled).toBeTruthy();
+
+                    // Item body is open
+                    let itemBodyDes = getAndExpectDebugElementByCss(
+                        compDe,
+                        'div#awg-graph-visualizer-construct-results > div.accordion-collapse',
+                        1,
+                        1,
+                        'open'
+                    );
+                    let itemBodyEl = itemBodyDes[0].nativeElement;
+
+                    expectToContain(itemBodyEl.classList, 'show');
+
+                    // Click header button
+                    click(btnEl as HTMLElement);
+                    detectChangesOnPush(fixture);
+
+                    // Item body does not close again
+                    itemBodyDes = getAndExpectDebugElementByCss(
+                        compDe,
+                        'div#awg-graph-visualizer-construct-results > div.accordion-collapse',
+                        1,
+                        1,
+                        'open'
+                    );
+                    itemBodyEl = itemBodyDes[0].nativeElement;
+
+                    expectToContain(itemBodyEl.classList, 'show');
+                });
+
+                describe('... should contain TwelveToneSpinnerComponent (stubbed) in item body while loading if ... ', () => {
+                    it('... queryResult$ is EMPTY', () => {
+                        // Mock empty observable
+                        component.queryResult$ = EMPTY;
+                        detectChangesOnPush(fixture);
+
+                        // Item body
+                        const bodyDes = getAndExpectDebugElementByCss(
+                            compDe,
+                            'div#awg-graph-visualizer-construct-results-collapse > div.accordion-body',
+                            1,
+                            1
+                        );
+
+                        getAndExpectDebugElementByDirective(bodyDes[0], TwelveToneSpinnerStubComponent, 1, 1);
+                    });
+
+                    it('... queryResult$ is undefined', () => {
+                        // Mock undefined response
+                        component.queryResult$ = observableOf(undefined);
+                        detectChangesOnPush(fixture);
+
+                        // Item body
+                        const bodyDes = getAndExpectDebugElementByCss(
+                            compDe,
+                            'div#awg-graph-visualizer-construct-results-collapse > div.accordion-body',
+                            1,
+                            1
+                        );
+
+                        getAndExpectDebugElementByDirective(bodyDes[0], TwelveToneSpinnerStubComponent, 1, 1);
+                    });
+
+                    it('... queryResult$ is null', () => {
+                        // Mock null response
+                        component.queryResult$ = observableOf(null);
+                        detectChangesOnPush(fixture);
+
+                        // Item body
+                        const bodyDes = getAndExpectDebugElementByCss(
+                            compDe,
+                            'div#awg-graph-visualizer-construct-results-collapse > div.accordion-body',
+                            1,
+                            1
+                        );
+
+                        getAndExpectDebugElementByDirective(bodyDes[0], TwelveToneSpinnerStubComponent, 1, 1);
+                    });
+                });
+
+                describe('... should contain item body with SparqlNoResultsStubComponent (stubbed) if ... ', () => {
+                    it('... queryResult is empty array', () => {
+                        // Mock empty array
+                        component.queryResult$ = observableOf([]);
+                        detectChangesOnPush(fixture);
+
+                        // Item body
+                        const bodyDes = getAndExpectDebugElementByCss(
+                            compDe,
+                            'div#awg-graph-visualizer-construct-results-collapse > div.accordion-body',
+                            1,
+                            1
+                        );
+
+                        getAndExpectDebugElementByDirective(bodyDes[0], SparqlNoResultsStubComponent, 1, 1);
+                    });
+
+                    it('... queryResult is undefined or empty string', () => {
+                        // Mock undefined response
+                        component.queryResult$ = observableOf([
+                            { subject: undefined, predicate: undefined, object: undefined },
+                        ]);
+                        detectChangesOnPush(fixture);
+
+                        // Item body
+                        const bodyDes = getAndExpectDebugElementByCss(
+                            compDe,
+                            'div#awg-graph-visualizer-construct-results-collapse > div.accordion-body',
+                            1,
+                            1
+                        );
+
+                        getAndExpectDebugElementByDirective(bodyDes[0], SparqlNoResultsStubComponent, 1, 1);
+
+                        // Mock empty response
+                        component.queryResult$ = observableOf([{ subject: '', predicate: '', object: '' }]);
+                        detectChangesOnPush(fixture);
+
+                        // Item body
+                        getAndExpectDebugElementByCss(
+                            compDe,
+                            'div#awg-graph-visualizer-construct-results-collapse > div.accordion-body',
+                            1,
+                            1
+                        );
+
+                        getAndExpectDebugElementByDirective(bodyDes[0], SparqlNoResultsStubComponent, 1, 1);
+                    });
+
+                    it('... queryResult.subject is undefined or empty string', () => {
+                        // Mock undefined response
+                        component.queryResult$ = observableOf([
+                            {
+                                subject: undefined,
+                                predicate: 'example:has',
+                                object: 'example:Success',
+                            },
+                        ]);
+                        detectChangesOnPush(fixture);
+
+                        // Item body
+                        const bodyDes = getAndExpectDebugElementByCss(
+                            compDe,
+                            'div#awg-graph-visualizer-construct-results-collapse > div.accordion-body',
+                            1,
+                            1
+                        );
+
+                        getAndExpectDebugElementByDirective(bodyDes[0], SparqlNoResultsStubComponent, 1, 1);
+
+                        // Mock empty response
+                        component.queryResult$ = observableOf([
+                            {
+                                subject: '',
+                                predicate: 'example:has',
+                                object: 'example:Success',
+                            },
+                        ]);
+                        detectChangesOnPush(fixture);
+
+                        // Item body
+                        getAndExpectDebugElementByCss(
+                            compDe,
+                            'div#awg-graph-visualizer-construct-results-collapse > div.accordion-body',
+                            1,
+                            1
+                        );
+
+                        getAndExpectDebugElementByDirective(bodyDes[0], SparqlNoResultsStubComponent, 1, 1);
+                    });
+
+                    it('... queryResult.predicate is undefined or empty string', () => {
+                        // Mock undefined response
+                        component.queryResult$ = observableOf([
+                            {
+                                subject: 'example:Test',
+                                predicate: undefined,
+                                object: 'example:Success',
+                            },
+                        ]);
+                        detectChangesOnPush(fixture);
+
+                        // Item body
+                        const bodyDes = getAndExpectDebugElementByCss(
+                            compDe,
+                            'div#awg-graph-visualizer-construct-results-collapse > div.accordion-body',
+                            1,
+                            1
+                        );
+
+                        getAndExpectDebugElementByDirective(bodyDes[0], SparqlNoResultsStubComponent, 1, 1);
+
+                        // Mock empty response
+                        component.queryResult$ = observableOf([
+                            {
+                                subject: 'example:Test',
+                                predicate: '',
+                                object: 'example:Success',
+                            },
+                        ]);
+                        detectChangesOnPush(fixture);
+
+                        // Item body
+                        getAndExpectDebugElementByCss(
+                            compDe,
+                            'div#awg-graph-visualizer-construct-results-collapse > div.accordion-body',
+                            1,
+                            1
+                        );
+
+                        getAndExpectDebugElementByDirective(bodyDes[0], SparqlNoResultsStubComponent, 1, 1);
+                    });
+
+                    it('... queryResult.object is undefined or empty string', () => {
+                        // Mock undefined response
+                        component.queryResult$ = observableOf([
+                            {
+                                subject: 'example:Test',
+                                predicate: 'example:has',
+                                object: undefined,
+                            },
+                        ]);
+                        detectChangesOnPush(fixture);
+
+                        // Item body
+                        const bodyDes = getAndExpectDebugElementByCss(
+                            compDe,
+                            'div#awg-graph-visualizer-construct-results-collapse > div.accordion-body',
+                            1,
+                            1
+                        );
+
+                        getAndExpectDebugElementByDirective(bodyDes[0], SparqlNoResultsStubComponent, 1, 1);
+
+                        // Mock empty response
+                        component.queryResult$ = observableOf([
+                            {
+                                subject: 'example:Test',
+                                predicate: 'example:has',
+                                object: '',
+                            },
+                        ]);
+                        detectChangesOnPush(fixture);
+
+                        // Item body
+                        getAndExpectDebugElementByCss(
+                            compDe,
+                            'div#awg-graph-visualizer-construct-results-collapse > div.accordion-body',
+                            1,
+                            1
+                        );
+
+                        getAndExpectDebugElementByDirective(bodyDes[0], SparqlNoResultsStubComponent, 1, 1);
+                    });
+                });
+
+                it('... should contain item body with ForceGraphComponent (stubbed) if results are available', () => {
+                    // Item body
+                    const bodyDes = getAndExpectDebugElementByCss(
+                        compDe,
+                        'div#awg-graph-visualizer-construct-results-collapse > div.accordion-body',
+                        1,
+                        1
+                    );
+
+                    // ForceGraph
+                    getAndExpectDebugElementByDirective(bodyDes[0], ForceGraphStubComponent, 1, 1);
+                });
+
+                it('... should pass down `queryResult` and `defaultForceGraphHeight` to forceGraph component', () => {
+                    const forceGraphDes = getAndExpectDebugElementByDirective(compDe, ForceGraphStubComponent, 1, 1);
+                    const forceGraphCmp = forceGraphDes[0].injector.get(
+                        ForceGraphStubComponent
+                    ) as ForceGraphStubComponent;
+
+                    expectToEqual(forceGraphCmp.queryResultTriples, [expectedQueryResult]);
+                    expectToBe(forceGraphCmp.height, expectedHeight);
+                });
+            });
+        });
+
+        describe('#isAccordionItemDisabled()', () => {
+            it('... should have a method `isAccordionItemDisabled`', () => {
+                expect(component.isAccordionItemDisabled).toBeDefined();
             });
 
-            it('... should not toggle panel body on click in fullscreen mode', () => {
+            it('... should be triggered from ngbAccordionItem', () => {
+                expectSpyCall(isAccordionItemDisabledSpy, 2);
+            });
+
+            it('... should return false if isFullscreen is false', () => {
+                expectToBe(component.isAccordionItemDisabled(), false);
+            });
+
+            it('... should return true if isFullscreen is true', () => {
+                // Set fullscreen flag to true
                 component.isFullscreen = true;
-                fixture.detectChanges();
 
-                // Header debug elements
-                const panelHeaderDes = getAndExpectDebugElementByCss(
-                    compDe,
-                    'div#awg-graph-visualizer-construct-results-header.accordion-header',
-                    1,
-                    1
-                );
+                expectToBe(component.isAccordionItemDisabled(), true);
+            });
+        });
 
-                // Button debug elements
-                const btnDes = getAndExpectDebugElementByCss(panelHeaderDes[0], 'button.accordion-button', 1, 1);
-                // Button native elements to click on
-                const btnEl = btnDes[0].nativeElement;
+        describe('#isQueryResultNotEmpty()', () => {
+            it('... should have a method `isQueryResultNotEmpty`', () => {
+                expect(component.isQueryResultNotEmpty).toBeDefined();
+            });
 
-                // Panel body does not close
-                getAndExpectDebugElementByCss(
-                    compDe,
-                    'div#awg-graph-visualizer-construct-results > div.accordion-body',
-                    1,
-                    1,
-                    'open'
-                );
+            it('... should be triggered from ngbAccordionBody', () => {
+                expectSpyCall(isQueryResultNotEmptySpy, 2, expectedQueryResult[0]);
+            });
 
-                // Click header button
-                click(btnEl as HTMLElement);
+            it('... should be triggered by change of queryResult', () => {
+                expectSpyCall(isQueryResultNotEmptySpy, 2, expectedQueryResult[0]);
+
+                // Mock another queryResult
+                const anotherQueryResult = {
+                    subject: 'example:AnotherTest',
+                    predicate: 'example:has',
+                    object: 'example:AnotherSuccess',
+                };
+                component.queryResult$ = observableOf([anotherQueryResult]);
                 detectChangesOnPush(fixture);
 
-                // Panel body does not close again
-                getAndExpectDebugElementByCss(
-                    compDe,
-                    'div#awg-graph-visualizer-construct-results > div.accordion-body',
-                    1,
-                    1,
-                    'open'
-                );
+                expectSpyCall(isQueryResultNotEmptySpy, 3, anotherQueryResult[0]);
             });
 
-            it('... should contain panel body with TwelveToneSpinnerComponent (stubbed) while loading', () => {
-                // Mock empty observable
-                component.queryResult$ = EMPTY;
+            describe('... should return false if ...', () => {
+                it('... queryResult is empty array', () => {
+                    expectSpyCall(isQueryResultNotEmptySpy, 2, expectedQueryResult[0]);
+
+                    // Mock empty response
+                    const emptyQueryResult = [];
+                    component.queryResult$ = observableOf(emptyQueryResult);
+                    detectChangesOnPush(fixture);
+
+                    expectSpyCall(isQueryResultNotEmptySpy, 3, emptyQueryResult);
+                    expectToBe(component.isQueryResultNotEmpty(emptyQueryResult), false);
+                });
+
+                it('... queryResult.subject is undefined or empty string', () => {
+                    expectSpyCall(isQueryResultNotEmptySpy, 2, expectedQueryResult[0]);
+
+                    // Mock undefined response
+                    const undefinedQueryResult = {
+                        subject: undefined,
+                        predicate: 'example:has',
+                        object: 'example:Success',
+                    };
+                    component.queryResult$ = observableOf([undefinedQueryResult]);
+                    detectChangesOnPush(fixture);
+
+                    expectSpyCall(isQueryResultNotEmptySpy, 3, undefinedQueryResult[0]);
+                    expectToBe(component.isQueryResultNotEmpty([undefinedQueryResult]), false);
+
+                    // Mock empty response
+                    const emptyQueryResult = {
+                        subject: '',
+                        predicate: 'example:has',
+                        object: 'example:Success',
+                    };
+                    component.queryResult$ = observableOf([emptyQueryResult]);
+                    detectChangesOnPush(fixture);
+
+                    expectSpyCall(isQueryResultNotEmptySpy, 5, emptyQueryResult[0]);
+                    expectToBe(component.isQueryResultNotEmpty([emptyQueryResult]), false);
+                });
+
+                it('... queryResult.predicate is undefined or empty string', () => {
+                    expectSpyCall(isQueryResultNotEmptySpy, 2, expectedQueryResult[0]);
+
+                    // Mock undefined response
+                    const undefinedQueryResult = {
+                        subject: 'example:Test',
+                        predicate: undefined,
+                        object: 'example:Success',
+                    };
+                    component.queryResult$ = observableOf([undefinedQueryResult]);
+                    detectChangesOnPush(fixture);
+
+                    expectSpyCall(isQueryResultNotEmptySpy, 3, undefinedQueryResult[0]);
+                    expectToBe(component.isQueryResultNotEmpty([undefinedQueryResult]), false);
+
+                    // Mock empty response
+                    const emptyQueryResult = {
+                        subject: 'example:Test',
+                        predicate: '',
+                        object: 'example:Success',
+                    };
+                    component.queryResult$ = observableOf([emptyQueryResult]);
+                    detectChangesOnPush(fixture);
+
+                    expectSpyCall(isQueryResultNotEmptySpy, 5, emptyQueryResult[0]);
+                    expectToBe(component.isQueryResultNotEmpty([emptyQueryResult]), false);
+                });
+
+                it('... queryResult.object is undefined or empty string', () => {
+                    expectSpyCall(isQueryResultNotEmptySpy, 2, expectedQueryResult[0]);
+
+                    // Mock undefined response
+                    const undefinedQueryResult = {
+                        subject: 'example:Test',
+                        predicate: 'example:has',
+                        object: undefined,
+                    };
+                    component.queryResult$ = observableOf([undefinedQueryResult]);
+                    detectChangesOnPush(fixture);
+
+                    expectSpyCall(isQueryResultNotEmptySpy, 3, undefinedQueryResult[0]);
+                    expectToBe(component.isQueryResultNotEmpty([undefinedQueryResult]), false);
+
+                    // Mock empty response
+                    const emptyQueryResult = {
+                        subject: 'example:Test',
+                        predicate: 'example:has',
+                        object: '',
+                    };
+                    component.queryResult$ = observableOf([emptyQueryResult]);
+                    detectChangesOnPush(fixture);
+
+                    expectSpyCall(isQueryResultNotEmptySpy, 5, emptyQueryResult[0]);
+                    expectToBe(component.isQueryResultNotEmpty([emptyQueryResult]), false);
+                });
+
+                it('... queryResult.subject, queryResult.predicate and queryResult.object are undefined or empty string', () => {
+                    expectSpyCall(isQueryResultNotEmptySpy, 2, expectedQueryResult[0]);
+
+                    // Mock undefined response
+                    const undefinedQueryResult = { subject: undefined, predicate: undefined, object: undefined };
+                    component.queryResult$ = observableOf([undefinedQueryResult]);
+                    detectChangesOnPush(fixture);
+
+                    expectSpyCall(isQueryResultNotEmptySpy, 3, undefinedQueryResult[0]);
+                    expectToBe(component.isQueryResultNotEmpty([undefinedQueryResult]), false);
+
+                    // Mock empty response
+                    const emptyQueryResult = { subject: '', predicate: '', object: '' };
+                    component.queryResult$ = observableOf([emptyQueryResult]);
+                    detectChangesOnPush(fixture);
+
+                    expectSpyCall(isQueryResultNotEmptySpy, 5, emptyQueryResult[0]);
+                    expectToBe(component.isQueryResultNotEmpty([emptyQueryResult]), false);
+                });
+            });
+
+            it('... should return true if queryResult is not empty', () => {
+                expectSpyCall(isQueryResultNotEmptySpy, 2, expectedQueryResult[0]);
+
+                // Mock non-empty response
+                const nonEmptyQueryResult = {
+                    subject: 'example:Test',
+                    predicate: 'example:has',
+                    object: 'example:Success',
+                };
+                component.queryResult$ = observableOf([nonEmptyQueryResult]);
                 detectChangesOnPush(fixture);
 
-                // Panel body
-                const bodyDes = getAndExpectDebugElementByCss(
-                    compDe,
-                    'div#awg-graph-visualizer-construct-results > div.accordion-body',
-                    1,
-                    1
-                );
-
-                getAndExpectDebugElementByDirective(bodyDes[0], TwelveToneSpinnerStubComponent, 1, 1);
-            });
-
-            it('... should contain panel body with SparqlNoResultsStubComponent (stubbed) if no results are available', () => {
-                // Mock empty response
-                component.queryResult$ = observableOf([]);
-                detectChangesOnPush(fixture);
-
-                // Panel body
-                const bodyDes = getAndExpectDebugElementByCss(
-                    compDe,
-                    'div#awg-graph-visualizer-construct-results > div.accordion-body',
-                    1,
-                    1
-                );
-
-                // SparqlNoResultsStubComponent
-                getAndExpectDebugElementByDirective(bodyDes[0], SparqlNoResultsStubComponent, 1, 1);
-            });
-
-            it('... should contain panel body with ForceGraphComponent (stubbed) if results are available', () => {
-                // Panel body
-                const bodyDes = getAndExpectDebugElementByCss(
-                    compDe,
-                    'div#awg-graph-visualizer-construct-results > div.accordion-body',
-                    1,
-                    1
-                );
-
-                // ForceGraph
-                getAndExpectDebugElementByDirective(bodyDes[0], ForceGraphStubComponent, 1, 1);
-            });
-
-            it('... should pass down `queryResult` and `defaultForceGraphHeight` to forceGraph component', () => {
-                const forceGraphDes = getAndExpectDebugElementByDirective(compDe, ForceGraphStubComponent, 1, 1);
-                const forceGraphCmp = forceGraphDes[0].injector.get(ForceGraphStubComponent) as ForceGraphStubComponent;
-
-                expect(forceGraphCmp.queryResultTriples).toBeDefined();
-                expect(forceGraphCmp.queryResultTriples)
-                    .withContext(`should equal ${expectedTriples}`)
-                    .toEqual(expectedTriples);
-
-                expect(forceGraphCmp.height).toBeDefined();
-                expect(forceGraphCmp.height).withContext(`should be: ${expectedHeight}`).toBe(expectedHeight);
+                expectSpyCall(isQueryResultNotEmptySpy, 3, nonEmptyQueryResult[0]);
+                expectToBe(component.isQueryResultNotEmpty([nonEmptyQueryResult]), true);
             });
         });
 
@@ -375,109 +1073,6 @@ describe('ConstructResultsComponent (DONE)', () => {
 
                 expectSpyCall(nodeClickSpy, 1, node);
                 expectSpyCall(emitClickedNodeRequestSpy, 1, node);
-            });
-        });
-
-        describe('#preventPanelCollapseOnFullscreen()', () => {
-            it('... should have a method `preventPanelCollapseOnFullscreen`', () => {
-                expect(component.preventPanelCollapseOnFullscreen).toBeDefined();
-            });
-
-            it('... should trigger on event from ngb-accordion', () => {
-                const accordionDes = getAndExpectDebugElementByDirective(compDe, NgbAccordion, 1, 1);
-                const accordionCmp = accordionDes[0].injector.get(NgbAccordion) as NgbAccordion;
-
-                const panelChangeEvent: NgbPanelChangeEvent = {
-                    panelId: 'panelChangeId',
-                    nextState: true,
-                    preventDefault: () => {
-                        // Intentional empty test override
-                    },
-                };
-
-                // Emit change event from accordion
-                accordionCmp.panelChange.emit(panelChangeEvent);
-
-                expectSpyCall(preventPanelCollapseOnFullscreenSpy, 1, panelChangeEvent);
-            });
-
-            it('... should not do anything if no $event is provided', () => {
-                const accordionDes = getAndExpectDebugElementByDirective(compDe, NgbAccordion, 1, 1);
-                const accordionCmp = accordionDes[0].injector.get(NgbAccordion) as NgbAccordion;
-
-                // Emit undefined change event from accordion
-                accordionCmp.panelChange.emit(undefined);
-
-                expectSpyCall(preventPanelCollapseOnFullscreenSpy, 1, undefined);
-            });
-
-            it('... should trigger $event.preventDefault() if isFullscreen == true && $event.nextState == false', () => {
-                const accordionDes = getAndExpectDebugElementByDirective(compDe, NgbAccordion, 1, 1);
-                const accordionCmp = accordionDes[0].injector.get(NgbAccordion) as NgbAccordion;
-
-                const panelChangeEvent: NgbPanelChangeEvent = {
-                    panelId: 'panelChangeId',
-                    nextState: false,
-                    preventDefault: () => {
-                        // Intentional empty test override
-                    },
-                };
-                const preventDefaultSpy: Spy = spyOn(panelChangeEvent, 'preventDefault').and.callThrough();
-
-                // Set fullscreen mode
-                component.isFullscreen = true;
-
-                // Emit change event from accordion
-                accordionCmp.panelChange.emit(panelChangeEvent);
-
-                expectSpyCall(preventPanelCollapseOnFullscreenSpy, 1, panelChangeEvent);
-                expectSpyCall(preventDefaultSpy, 1, undefined);
-            });
-
-            it('... should not trigger $event.preventDefault() if $event.nextState == true', () => {
-                const accordionDes = getAndExpectDebugElementByDirective(compDe, NgbAccordion, 1, 1);
-                const accordionCmp = accordionDes[0].injector.get(NgbAccordion) as NgbAccordion;
-
-                const panelChangeEvent: NgbPanelChangeEvent = {
-                    panelId: 'panelChangeId',
-                    nextState: true,
-                    preventDefault: () => {
-                        // Intentional empty test override
-                    },
-                };
-                const preventDefaultSpy: Spy = spyOn(panelChangeEvent, 'preventDefault').and.callThrough();
-
-                // Set fullscreen mode
-                component.isFullscreen = true;
-
-                // Emit change event from accordion
-                accordionCmp.panelChange.emit(panelChangeEvent);
-
-                expectSpyCall(preventPanelCollapseOnFullscreenSpy, 1, panelChangeEvent);
-                expectSpyCall(preventDefaultSpy, 0, undefined);
-            });
-
-            it('... should not trigger $event.preventDefault() if isFullscreen == false', () => {
-                const accordionDes = getAndExpectDebugElementByDirective(compDe, NgbAccordion, 1, 1);
-                const accordionCmp = accordionDes[0].injector.get(NgbAccordion) as NgbAccordion;
-
-                const panelChangeEvent: NgbPanelChangeEvent = {
-                    panelId: 'panelChangeId',
-                    nextState: false,
-                    preventDefault: () => {
-                        // Intentional empty test override
-                    },
-                };
-                const preventDefaultSpy: Spy = spyOn(panelChangeEvent, 'preventDefault').and.callThrough();
-
-                // Unset fullscreen mode
-                component.isFullscreen = false;
-
-                // Emit change event from accordion
-                accordionCmp.panelChange.emit(panelChangeEvent);
-
-                expectSpyCall(preventPanelCollapseOnFullscreenSpy, 1, panelChangeEvent);
-                expectSpyCall(preventDefaultSpy, 0, undefined);
             });
         });
     });
