@@ -1,16 +1,18 @@
 import { Component, DebugElement, EventEmitter, Input, NgModule, Output } from '@angular/core';
 import { ComponentFixture, TestBed, waitForAsync } from '@angular/core/testing';
 
-import { EMPTY, Observable, of as observableOf } from 'rxjs';
+import { EMPTY, Observable, lastValueFrom, of as observableOf } from 'rxjs';
 import Spy = jasmine.Spy;
 
-import { NgbAccordion, NgbAccordionModule, NgbConfig, NgbPanelChangeEvent } from '@ng-bootstrap/ng-bootstrap';
+import { NgbAccordionDirective, NgbAccordionModule, NgbConfig } from '@ng-bootstrap/ng-bootstrap';
 
 import { click } from '@testing/click-helper';
 import { detectChangesOnPush } from '@testing/detect-changes-on-push-helper';
-
 import {
     expectSpyCall,
+    expectToBe,
+    expectToContain,
+    expectToEqual,
     getAndExpectDebugElementByCss,
     getAndExpectDebugElementByDirective,
 } from '@testing/expect-helper';
@@ -42,10 +44,10 @@ describe('SelectResultsComponent (DONE)', () => {
     let expectedQueryTime: number;
     let expectedIsFullscreen: boolean;
 
-    let tableClickSpy: Spy;
     let emitClickedTableRequestSpy: Spy;
-    let isNotEmptySpy: Spy;
-    let preventPanelCollapseOnFullscreenSpy: Spy;
+    let isAccordionItemDisabledSpy: Spy;
+    let isQueryResultNotEmptySpy: Spy;
+    let tableClickSpy: Spy;
 
     // Global NgbConfigModule
     @NgModule({ imports: [NgbAccordionModule], exports: [NgbAccordionModule] })
@@ -58,7 +60,7 @@ describe('SelectResultsComponent (DONE)', () => {
 
     beforeEach(waitForAsync(() => {
         TestBed.configureTestingModule({
-            imports: [NgbAccordionWithConfigModule],
+            imports: [NgbAccordionWithConfigModule, NgbAccordionDirective],
             declarations: [
                 SelectResultsComponent,
                 SparqlNoResultsStubComponent,
@@ -89,10 +91,10 @@ describe('SelectResultsComponent (DONE)', () => {
         // Spies on component functions
         // `.and.callThrough` will track the spy down the nested describes, see
         // https://jasmine.github.io/2.0/introduction.html#section-Spies:_%3Ccode%3Eand.callThrough%3C/code%3E
-        tableClickSpy = spyOn(component, 'onTableNodeClick').and.callThrough();
         emitClickedTableRequestSpy = spyOn(component.clickedTableRequest, 'emit').and.callThrough();
-        isNotEmptySpy = spyOn(component, 'isNotEmpty').and.callThrough();
-        preventPanelCollapseOnFullscreenSpy = spyOn(component, 'preventPanelCollapseOnFullscreen').and.callThrough();
+        isAccordionItemDisabledSpy = spyOn(component, 'isAccordionItemDisabled').and.callThrough();
+        isQueryResultNotEmptySpy = spyOn(component, 'isQueryResultNotEmpty').and.callThrough();
+        tableClickSpy = spyOn(component, 'onTableNodeClick').and.callThrough();
     });
 
     it('... should create', () => {
@@ -113,12 +115,9 @@ describe('SelectResultsComponent (DONE)', () => {
         });
 
         describe('VIEW', () => {
-            it('... should contain one ngb-accordion without panel (div.accordion-item) yet', () => {
-                // Ngb-accordion debug element
-                const accordionDes = getAndExpectDebugElementByCss(compDe, 'ngb-accordion', 1, 1);
-
-                // Panel
-                getAndExpectDebugElementByCss(accordionDes[0], 'div.accordion-item', 0, 0, 'yet');
+            it('... should contain no div.accordion yet', () => {
+                // Div.accordion debug element
+                getAndExpectDebugElementByCss(compDe, 'div.accordion', 0, 0);
             });
         });
     });
@@ -134,218 +133,666 @@ describe('SelectResultsComponent (DONE)', () => {
             fixture.detectChanges();
         });
 
-        it('... should have `queryResult` input', () => {
-            expect(component.queryResult$).toBeDefined();
-            expect(component.queryResult$)
-                .withContext(`should equal ${expectedQueryResult$}`)
-                .toEqual(expectedQueryResult$);
-        });
+        it('... should have `queryResult` input', waitForAsync(() => {
+            expectToEqual(component.queryResult$, expectedQueryResult$);
+            expectAsync(lastValueFrom(component.queryResult$)).toBeResolved();
+            expectAsync(lastValueFrom(component.queryResult$))
+                .withContext(`should be resolved to ${expectedQueryResult}}`)
+                .toBeResolvedTo(expectedQueryResult);
+        }));
 
         it('... should have `queryTime` input', () => {
-            expect(component.queryTime).toBeDefined();
-            expect(component.queryTime).withContext(`should equal ${expectedQueryTime}`).toEqual(expectedQueryTime);
+            expectToBe(component.queryTime, expectedQueryTime);
         });
 
         it('... should have `isFullscreen` input', () => {
-            expect(component.isFullscreen).toBeDefined();
-            expect(component.isFullscreen)
-                .withContext(`should equal ${expectedIsFullscreen}`)
-                .toEqual(expectedIsFullscreen);
+            expectToBe(component.isFullscreen, expectedIsFullscreen);
         });
 
         describe('VIEW', () => {
-            it('... should contain one ngb-accordion with panel (div.accordion-item) header and body', () => {
-                // Ngb-accordion debug element
-                const accordionDes = getAndExpectDebugElementByCss(compDe, 'ngb-accordion', 1, 1);
-
-                // Panel (div.card)
-                const panelDes = getAndExpectDebugElementByCss(accordionDes[0], 'div.accordion-item', 1, 1); // Panel (div.card)
-                // Header
-                getAndExpectDebugElementByCss(
-                    panelDes[0],
-                    'div#awg-graph-visualizer-select-results-header.accordion-header',
-                    1,
-                    1
-                ); // Panel (div.card)
-                // Body
-                getAndExpectDebugElementByCss(
-                    panelDes[0],
-                    'div#awg-graph-visualizer-select-results > div.accordion-body',
-                    1,
-                    1
-                );
+            it('... should contain one div.accordion', () => {
+                // NgbAccordion debug element
+                getAndExpectDebugElementByCss(compDe, 'div.accordion', 1, 1);
             });
 
-            it('... should display panel header button', () => {
-                // Panel header button
-                const btnDes = getAndExpectDebugElementByCss(
-                    compDe,
-                    'div#awg-graph-visualizer-select-results-header > button.accordion-button',
-                    1,
-                    1
-                );
+            describe('not in fullscreen mode', () => {
+                it('... should contain one div.accordion-item with header and open body in div.accordion', () => {
+                    // NgbAccordion debug element
+                    const accordionDes = getAndExpectDebugElementByCss(compDe, 'div.accordion', 1, 1);
 
-                const btnEl = btnDes[0].nativeElement;
+                    // Div.accordion-item
+                    const itemDes = getAndExpectDebugElementByCss(
+                        accordionDes[0],
+                        'div#awg-graph-visualizer-select-results.accordion-item',
+                        1,
+                        1
+                    );
+                    // Header (div.accordion-header)
+                    getAndExpectDebugElementByCss(
+                        itemDes[0],
+                        'div#awg-graph-visualizer-select-results > div.accordion-header',
+                        1,
+                        1
+                    );
 
-                // Check button content
-                expect(btnEl.textContent).toBeTruthy();
-                expect(btnEl.textContent).withContext('should contain Resultat').toContain('Resultat');
+                    // Body open (div.accordion-collapse)
+                    const itemBodyDes = getAndExpectDebugElementByCss(
+                        itemDes[0],
+                        'div#awg-graph-visualizer-select-results-collapse',
+                        1,
+                        1
+                    );
+                    const itemBodyEl = itemBodyDes[0].nativeElement;
+
+                    expectToContain(itemBodyEl.classList, 'show');
+                });
+
+                it('... should display item header button', () => {
+                    // Header debug elements
+                    const itemHeaderDes = getAndExpectDebugElementByCss(
+                        compDe,
+                        'div#awg-graph-visualizer-select-results > div.accordion-header',
+                        1,
+                        1
+                    );
+
+                    // Item header button
+                    const btnDes = getAndExpectDebugElementByCss(itemHeaderDes[0], 'button.accordion-button', 1, 1);
+                    const btnEl = btnDes[0].nativeElement;
+
+                    // Check button content
+                    expectToBe(btnEl.textContent, 'Resultat');
+                });
+
+                it('... should toggle item body on click', () => {
+                    // Header debug elements
+                    const itemHeaderDes = getAndExpectDebugElementByCss(
+                        compDe,
+                        'div#awg-graph-visualizer-select-results > div.accordion-header',
+                        1,
+                        1
+                    );
+
+                    // Button debug elements
+                    const btnDes = getAndExpectDebugElementByCss(itemHeaderDes[0], 'button.accordion-button', 1, 1);
+                    // Button native elements to click on
+                    const btnEl = btnDes[0].nativeElement;
+
+                    // Item body is open
+                    let itemBodyDes = getAndExpectDebugElementByCss(
+                        compDe,
+                        'div#awg-graph-visualizer-select-results > div.accordion-collapse',
+                        1,
+                        1,
+                        'open'
+                    );
+                    let itemBodyEl = itemBodyDes[0].nativeElement;
+
+                    expectToContain(itemBodyEl.classList, 'show');
+
+                    // Click header button
+                    click(btnEl as HTMLElement);
+                    detectChangesOnPush(fixture);
+
+                    // Item body is collapsed
+                    itemBodyDes = getAndExpectDebugElementByCss(
+                        compDe,
+                        'div#awg-graph-visualizer-select-results > div.accordion-collapse',
+                        1,
+                        1,
+                        'collapsed'
+                    );
+                    itemBodyEl = itemBodyDes[0].nativeElement;
+
+                    expectToContain(itemBodyEl.classList, 'collapse');
+
+                    // Click header button
+                    click(btnEl as HTMLElement);
+                    detectChangesOnPush(fixture);
+
+                    // Item body is open again
+                    itemBodyDes = getAndExpectDebugElementByCss(
+                        compDe,
+                        'div#awg-graph-visualizer-select-results > div.accordion-collapse',
+                        1,
+                        1,
+                        'open'
+                    );
+                    itemBodyEl = itemBodyDes[0].nativeElement;
+
+                    expectToContain(itemBodyEl.classList, 'show');
+                });
+
+                describe('... should contain TwelveToneSpinnerComponent (stubbed) in item body while loading if ... ', () => {
+                    it('... queryResult$ is EMPTY', () => {
+                        // Mock empty observable
+                        component.queryResult$ = EMPTY;
+                        detectChangesOnPush(fixture);
+
+                        // Item body
+                        const bodyDes = getAndExpectDebugElementByCss(
+                            compDe,
+                            'div#awg-graph-visualizer-select-results-collapse > div.accordion-body',
+                            1,
+                            1
+                        );
+
+                        getAndExpectDebugElementByDirective(bodyDes[0], TwelveToneSpinnerStubComponent, 1, 1);
+                    });
+
+                    it('... queryResult$ is undefined', () => {
+                        // Mock undefined response
+                        component.queryResult$ = undefined;
+                        detectChangesOnPush(fixture);
+
+                        // Item body
+                        const bodyDes = getAndExpectDebugElementByCss(
+                            compDe,
+                            'div#awg-graph-visualizer-select-results-collapse > div.accordion-body',
+                            1,
+                            1
+                        );
+
+                        getAndExpectDebugElementByDirective(bodyDes[0], TwelveToneSpinnerStubComponent, 1, 1);
+                    });
+
+                    it('... queryResult$ is null', () => {
+                        // Mock null response
+                        component.queryResult$ = null;
+                        detectChangesOnPush(fixture);
+
+                        // Item body
+                        const bodyDes = getAndExpectDebugElementByCss(
+                            compDe,
+                            'div#awg-graph-visualizer-select-results-collapse > div.accordion-body',
+                            1,
+                            1
+                        );
+
+                        getAndExpectDebugElementByDirective(bodyDes[0], TwelveToneSpinnerStubComponent, 1, 1);
+                    });
+                });
+
+                describe('... should contain item body with SparqlNoResultsStubComponent (stubbed) if ... ', () => {
+                    it('... no results are available', () => {
+                        // Mock empty response
+                        component.queryResult$ = observableOf({ head: { vars: [] }, body: { bindings: [] } });
+                        detectChangesOnPush(fixture);
+
+                        // Item body
+                        const bodyDes = getAndExpectDebugElementByCss(
+                            compDe,
+                            'div#awg-graph-visualizer-select-results-collapse > div.accordion-body',
+                            1,
+                            1
+                        );
+
+                        // SparqlNoResultsStubComponent
+                        getAndExpectDebugElementByDirective(bodyDes[0], SparqlNoResultsStubComponent, 1, 1);
+                    });
+
+                    it('... queryResult.head is undefined', () => {
+                        // Mock undefined response
+                        const undefinedQueryResult = { head: undefined, body: { bindings: [{ test: 'Test' }] } };
+                        component.queryResult$ = observableOf(undefinedQueryResult);
+                        detectChangesOnPush(fixture);
+
+                        // Item body
+                        const bodyDes = getAndExpectDebugElementByCss(
+                            compDe,
+                            'div#awg-graph-visualizer-select-results-collapse > div.accordion-body',
+                            1,
+                            1
+                        );
+
+                        // SparqlNoResultsStubComponent
+                        getAndExpectDebugElementByDirective(bodyDes[0], SparqlNoResultsStubComponent, 1, 1);
+                    });
+
+                    it('... queryResult.body is undefined', () => {
+                        // Mock undefined response
+                        const undefinedQueryResult = { head: { vars: ['Test'] }, body: undefined };
+                        component.queryResult$ = observableOf(undefinedQueryResult);
+                        detectChangesOnPush(fixture);
+
+                        // Item body
+                        const bodyDes = getAndExpectDebugElementByCss(
+                            compDe,
+                            'div#awg-graph-visualizer-select-results-collapse > div.accordion-body',
+                            1,
+                            1
+                        );
+
+                        // SparqlNoResultsStubComponent
+                        getAndExpectDebugElementByDirective(bodyDes[0], SparqlNoResultsStubComponent, 1, 1);
+                    });
+
+                    it('... queryResult.head and queryResult.body are undefined', () => {
+                        // Mock undefined response
+                        const undefinedQueryResult = { head: undefined, body: undefined };
+                        component.queryResult$ = observableOf(undefinedQueryResult);
+                        detectChangesOnPush(fixture);
+
+                        // Item body
+                        const body = getAndExpectDebugElementByCss(
+                            compDe,
+                            'div#awg-graph-visualizer-select-results-collapse > div.accordion-body',
+                            1,
+                            1
+                        );
+
+                        // SparqlNoResultsStubComponent
+                        getAndExpectDebugElementByDirective(body[0], SparqlNoResultsStubComponent, 1, 1);
+                    });
+                });
+
+                it('... should contain item body with SparqlTableComponent (stubbed) if results are available', () => {
+                    // Item body
+                    const bodyDes = getAndExpectDebugElementByCss(
+                        compDe,
+                        'div#awg-graph-visualizer-select-results-collapse > div.accordion-body',
+                        1,
+                        1
+                    );
+
+                    // SparqlTable
+                    getAndExpectDebugElementByDirective(bodyDes[0], SparqlTableStubComponent, 1, 1);
+                });
+
+                it('... should pass down `queryResult` and `queryTime` to sparqlTable component', () => {
+                    const sparqlTableDes = getAndExpectDebugElementByDirective(compDe, SparqlTableStubComponent, 1, 1);
+                    const sparqlTableCmp = sparqlTableDes[0].injector.get(
+                        SparqlTableStubComponent
+                    ) as SparqlTableStubComponent;
+
+                    expectToEqual(sparqlTableCmp.queryResult, expectedQueryResult);
+                    expectToBe(sparqlTableCmp.queryTime, expectedQueryTime);
+                });
             });
 
-            it('... should toggle panel body on click when not in fullscreen mode', () => {
-                component.isFullscreen = false;
-                fixture.detectChanges();
+            describe('in fullscreen mode', () => {
+                beforeEach(() => {
+                    // Set fullscreen mode
+                    component.isFullscreen = true;
+                    detectChangesOnPush(fixture);
+                });
 
-                // Header debug elements
-                const panelHeaderDes = getAndExpectDebugElementByCss(
-                    compDe,
-                    'div#awg-graph-visualizer-select-results-header.accordion-header',
-                    1,
-                    1
-                );
+                it('... should contain one div.accordion-item with header and open body in div.accordion', () => {
+                    // NgbAccordion debug element
+                    const accordionDes = getAndExpectDebugElementByCss(compDe, 'div.accordion', 1, 1);
 
-                // Button debug elements
-                const btnDes = getAndExpectDebugElementByCss(panelHeaderDes[0], 'button.accordion-button', 1, 1);
-                // Button native elements to click on
-                const btnEl = btnDes[0].nativeElement;
+                    // Div.accordion-item
+                    const itemDes = getAndExpectDebugElementByCss(
+                        accordionDes[0],
+                        'div#awg-graph-visualizer-select-results.accordion-item',
+                        1,
+                        1
+                    );
+                    // Header (div.accordion-header)
+                    getAndExpectDebugElementByCss(
+                        itemDes[0],
+                        'div#awg-graph-visualizer-select-results > div.accordion-header',
+                        1,
+                        1
+                    );
 
-                // Panel body is open
-                getAndExpectDebugElementByCss(
-                    compDe,
-                    'div#awg-graph-visualizer-select-results > div.accordion-body',
-                    1,
-                    1,
-                    'open'
-                );
+                    // Body open (div.accordion-collapse)
+                    const itemBodyDes = getAndExpectDebugElementByCss(
+                        itemDes[0],
+                        'div#awg-graph-visualizer-select-results-collapse',
+                        1,
+                        1
+                    );
+                    const itemBodyEl = itemBodyDes[0].nativeElement;
 
-                // Click header button
-                click(btnEl as HTMLElement);
-                detectChangesOnPush(fixture);
+                    expectToContain(itemBodyEl.classList, 'show');
+                });
 
-                // Panel is open
-                getAndExpectDebugElementByCss(
-                    compDe,
-                    'div#awg-graph-visualizer-select-results > div.accordion-body',
-                    0,
-                    0,
-                    'collapsed'
-                );
+                it('... should display item header button', () => {
+                    // Header debug elements
+                    const itemHeaderDes = getAndExpectDebugElementByCss(
+                        compDe,
+                        'div#awg-graph-visualizer-select-results > div.accordion-header',
+                        1,
+                        1
+                    );
 
-                // Click header button
-                click(btnEl as HTMLElement);
-                detectChangesOnPush(fixture);
+                    // Item header button
+                    const btnDes = getAndExpectDebugElementByCss(itemHeaderDes[0], 'button.accordion-button', 1, 1);
+                    const btnEl = btnDes[0].nativeElement;
 
-                getAndExpectDebugElementByCss(
-                    // Panel body is closed again
-                    compDe,
-                    'div#awg-graph-visualizer-select-results > div.accordion-body',
-                    1,
-                    1,
-                    'open'
-                );
+                    // Check button content
+                    expectToBe(btnEl.textContent, 'Resultat');
+                });
+
+                it('... should not toggle item body on click', () => {
+                    // Header debug elements
+                    const itemHeaderDes = getAndExpectDebugElementByCss(
+                        compDe,
+                        'div#awg-graph-visualizer-select-results > div.accordion-header',
+                        1,
+                        1
+                    );
+
+                    // Button debug elements
+                    const btnDes = getAndExpectDebugElementByCss(itemHeaderDes[0], 'button.accordion-button', 1, 1);
+                    // Button native elements to click on
+                    const btnEl = btnDes[0].nativeElement;
+
+                    expect(btnEl.disabled).toBeTruthy();
+
+                    // Item body is open
+                    let itemBodyDes = getAndExpectDebugElementByCss(
+                        compDe,
+                        'div#awg-graph-visualizer-select-results > div.accordion-collapse',
+                        1,
+                        1,
+                        'open'
+                    );
+                    let itemBodyEl = itemBodyDes[0].nativeElement;
+
+                    expectToContain(itemBodyEl.classList, 'show');
+
+                    // Click header button
+                    click(btnEl as HTMLElement);
+                    detectChangesOnPush(fixture);
+
+                    // Item body does not close again
+                    itemBodyDes = getAndExpectDebugElementByCss(
+                        compDe,
+                        'div#awg-graph-visualizer-select-results > div.accordion-collapse',
+                        1,
+                        1,
+                        'open'
+                    );
+                    itemBodyEl = itemBodyDes[0].nativeElement;
+
+                    expectToContain(itemBodyEl.classList, 'show');
+                });
+
+                describe('... should contain TwelveToneSpinnerComponent (stubbed) in item body while loading if ... ', () => {
+                    it('... queryResult$ is EMPTY', () => {
+                        // Mock empty observable
+                        component.queryResult$ = EMPTY;
+                        detectChangesOnPush(fixture);
+
+                        // Item body
+                        const bodyDes = getAndExpectDebugElementByCss(
+                            compDe,
+                            'div#awg-graph-visualizer-select-results-collapse > div.accordion-body',
+                            1,
+                            1
+                        );
+
+                        getAndExpectDebugElementByDirective(bodyDes[0], TwelveToneSpinnerStubComponent, 1, 1);
+                    });
+
+                    it('... queryResult$ is undefined', () => {
+                        // Mock undefined response
+                        component.queryResult$ = undefined;
+                        detectChangesOnPush(fixture);
+
+                        // Item body
+                        const bodyDes = getAndExpectDebugElementByCss(
+                            compDe,
+                            'div#awg-graph-visualizer-select-results-collapse > div.accordion-body',
+                            1,
+                            1
+                        );
+
+                        getAndExpectDebugElementByDirective(bodyDes[0], TwelveToneSpinnerStubComponent, 1, 1);
+                    });
+
+                    it('... queryResult$ is null', () => {
+                        // Mock null response
+                        component.queryResult$ = null;
+                        detectChangesOnPush(fixture);
+
+                        // Item body
+                        const bodyDes = getAndExpectDebugElementByCss(
+                            compDe,
+                            'div#awg-graph-visualizer-select-results-collapse > div.accordion-body',
+                            1,
+                            1
+                        );
+
+                        getAndExpectDebugElementByDirective(bodyDes[0], TwelveToneSpinnerStubComponent, 1, 1);
+                    });
+                });
+
+                describe('... should contain item body with SparqlNoResultsStubComponent (stubbed) if ... ', () => {
+                    it('... no results are available', () => {
+                        // Mock empty response
+                        component.queryResult$ = observableOf({ head: { vars: [] }, body: { bindings: [] } });
+                        detectChangesOnPush(fixture);
+
+                        // Item body
+                        const bodyDes = getAndExpectDebugElementByCss(
+                            compDe,
+                            'div#awg-graph-visualizer-select-results-collapse > div.accordion-body',
+                            1,
+                            1
+                        );
+
+                        // SparqlNoResultsStubComponent
+                        getAndExpectDebugElementByDirective(bodyDes[0], SparqlNoResultsStubComponent, 1, 1);
+                    });
+
+                    it('... queryResult.head is undefined', () => {
+                        // Mock undefined response
+                        const undefinedQueryResult = { head: undefined, body: { bindings: [{ test: 'Test' }] } };
+                        component.queryResult$ = observableOf(undefinedQueryResult);
+                        detectChangesOnPush(fixture);
+
+                        // Item body
+                        const bodyDes = getAndExpectDebugElementByCss(
+                            compDe,
+                            'div#awg-graph-visualizer-select-results-collapse > div.accordion-body',
+                            1,
+                            1
+                        );
+
+                        // SparqlNoResultsStubComponent
+                        getAndExpectDebugElementByDirective(bodyDes[0], SparqlNoResultsStubComponent, 1, 1);
+                    });
+
+                    it('... queryResult.body is undefined', () => {
+                        // Mock undefined response
+                        const undefinedQueryResult = { head: { vars: ['Test'] }, body: undefined };
+                        component.queryResult$ = observableOf(undefinedQueryResult);
+                        detectChangesOnPush(fixture);
+
+                        // Item body
+                        const bodyDes = getAndExpectDebugElementByCss(
+                            compDe,
+                            'div#awg-graph-visualizer-select-results-collapse > div.accordion-body',
+                            1,
+                            1
+                        );
+
+                        // SparqlNoResultsStubComponent
+                        getAndExpectDebugElementByDirective(bodyDes[0], SparqlNoResultsStubComponent, 1, 1);
+                    });
+
+                    it('... queryResult.head and queryResult.body are undefined', () => {
+                        // Mock undefined response
+                        const undefinedQueryResult = { head: undefined, body: undefined };
+                        component.queryResult$ = observableOf(undefinedQueryResult);
+                        detectChangesOnPush(fixture);
+
+                        // Item body
+                        const body = getAndExpectDebugElementByCss(
+                            compDe,
+                            'div#awg-graph-visualizer-select-results-collapse > div.accordion-body',
+                            1,
+                            1
+                        );
+
+                        // SparqlNoResultsStubComponent
+                        getAndExpectDebugElementByDirective(body[0], SparqlNoResultsStubComponent, 1, 1);
+                    });
+                });
+
+                it('... should contain item body with SparqlTableComponent (stubbed) if results are available', () => {
+                    // Item body
+                    const bodyDes = getAndExpectDebugElementByCss(
+                        compDe,
+                        'div#awg-graph-visualizer-select-results-collapse > div.accordion-body',
+                        1,
+                        1
+                    );
+
+                    // SparqlTable
+                    getAndExpectDebugElementByDirective(bodyDes[0], SparqlTableStubComponent, 1, 1);
+                });
+
+                it('... should pass down `queryResult` and `queryTime` to sparqlTable component', () => {
+                    const sparqlTableDes = getAndExpectDebugElementByDirective(compDe, SparqlTableStubComponent, 1, 1);
+                    const sparqlTableCmp = sparqlTableDes[0].injector.get(
+                        SparqlTableStubComponent
+                    ) as SparqlTableStubComponent;
+
+                    expectToEqual(sparqlTableCmp.queryResult, expectedQueryResult);
+                    expectToBe(sparqlTableCmp.queryTime, expectedQueryTime);
+                });
+            });
+        });
+
+        describe('#isAccordionItemDisabled()', () => {
+            it('... should have a method `isAccordionItemDisabled`', () => {
+                expect(component.isAccordionItemDisabled).toBeDefined();
             });
 
-            it('... should not toggle panel body on click in fullscreen mode', () => {
+            it('... should be triggered from ngbAccordionItem', () => {
+                expectSpyCall(isAccordionItemDisabledSpy, 2);
+            });
+
+            it('... should return false if isFullscreen is false', () => {
+                expectToBe(component.isAccordionItemDisabled(), false);
+            });
+
+            it('... should return true if isFullscreen is true', () => {
+                // Set fullscreen flag to true
                 component.isFullscreen = true;
-                fixture.detectChanges();
 
-                // Header debug elements
-                const panelHeaderDes = getAndExpectDebugElementByCss(
-                    compDe,
-                    'div#awg-graph-visualizer-select-results-header.accordion-header',
-                    1,
-                    1
-                );
+                expectToBe(component.isAccordionItemDisabled(), true);
+            });
+        });
 
-                // Button debug elements
-                const btnDes = getAndExpectDebugElementByCss(panelHeaderDes[0], 'button.accordion-button', 1, 1);
-                // Button native elements to click on
-                const btnEl = btnDes[0].nativeElement;
+        describe('#isQueryResultNotEmpty()', () => {
+            it('... should have a method `isQueryResultNotEmpty`', () => {
+                expect(component.isQueryResultNotEmpty).toBeDefined();
+            });
 
-                // Panel body does not close
-                getAndExpectDebugElementByCss(
-                    compDe,
-                    'div#awg-graph-visualizer-select-results > div.accordion-body',
-                    1,
-                    1,
-                    'open'
-                );
+            it('... should be triggered from ngbAccordionBody', () => {
+                expectSpyCall(isQueryResultNotEmptySpy, 2, expectedQueryResult[0]);
+            });
 
-                // Click header button
-                click(btnEl as HTMLElement);
+            it('... should be triggered by change of queryResult', () => {
+                expectSpyCall(isQueryResultNotEmptySpy, 2, expectedQueryResult);
+
+                // Mock another queryResult
+                const queryResult = {
+                    head: { vars: ['AnotherTestHeader'] },
+                    body: { bindings: [{ testKey: 'AnotherTestValue' }] },
+                };
+                component.queryResult$ = observableOf(queryResult);
+
                 detectChangesOnPush(fixture);
 
-                // Panel body does not close again
-                getAndExpectDebugElementByCss(
-                    compDe,
-                    'div#awg-graph-visualizer-select-results > div.accordion-body',
-                    1,
-                    1,
-                    'open'
-                );
+                expectSpyCall(isQueryResultNotEmptySpy, 3, queryResult);
             });
 
-            it('... should contain panel body with TwelveToneSpinnerComponent (stubbed) while loading', () => {
-                // Mock empty observable
-                component.queryResult$ = EMPTY;
+            describe('... should return false if ...', () => {
+                it('... queryResult.head is undefined', () => {
+                    expectSpyCall(isQueryResultNotEmptySpy, 2, expectedQueryResult);
+
+                    // Mock undefined response
+                    const undefinedQueryResult = { head: undefined, body: { bindings: [{ test: 'Test' }] } };
+                    component.queryResult$ = observableOf(undefinedQueryResult);
+                    detectChangesOnPush(fixture);
+
+                    expectSpyCall(isQueryResultNotEmptySpy, 3, undefinedQueryResult);
+                    expectToBe(component.isQueryResultNotEmpty(undefinedQueryResult), false);
+                });
+
+                it('... queryResult.body is undefined', () => {
+                    expectSpyCall(isQueryResultNotEmptySpy, 2, expectedQueryResult);
+
+                    // Mock undefined response
+                    const undefinedQueryResult = { head: { vars: ['Test'] }, body: undefined };
+                    component.queryResult$ = observableOf(undefinedQueryResult);
+                    detectChangesOnPush(fixture);
+
+                    expectSpyCall(isQueryResultNotEmptySpy, 3, undefinedQueryResult);
+                    expectToBe(component.isQueryResultNotEmpty(undefinedQueryResult), false);
+                });
+
+                it('... queryResult.head and queryResult.body are undefined', () => {
+                    expectSpyCall(isQueryResultNotEmptySpy, 2, expectedQueryResult);
+
+                    // Mock undefined response
+                    const undefinedQueryResult = { head: undefined, body: undefined };
+                    component.queryResult$ = observableOf(undefinedQueryResult);
+                    detectChangesOnPush(fixture);
+
+                    expectSpyCall(isQueryResultNotEmptySpy, 3, undefinedQueryResult);
+                    expectToBe(component.isQueryResultNotEmpty(undefinedQueryResult), false);
+                });
+
+                it('... queryResult.head.vars is empty array', () => {
+                    expectSpyCall(isQueryResultNotEmptySpy, 2, expectedQueryResult);
+
+                    // Mock empty response
+                    const emptyQueryResult = { head: { vars: [] }, body: { bindings: [{ testKey: 'TestValue' }] } };
+                    component.queryResult$ = observableOf(emptyQueryResult);
+
+                    detectChangesOnPush(fixture);
+
+                    expectSpyCall(isQueryResultNotEmptySpy, 3, emptyQueryResult);
+                    expectToBe(component.isQueryResultNotEmpty(emptyQueryResult), false);
+                });
+
+                it('... queryResult.body.bindings is empty array', () => {
+                    expectSpyCall(isQueryResultNotEmptySpy, 2, expectedQueryResult);
+
+                    // Mock empty response
+                    const emptyQueryResult = { head: { vars: ['TestHeader'] }, body: { bindings: [] } };
+                    component.queryResult$ = observableOf(emptyQueryResult);
+
+                    detectChangesOnPush(fixture);
+
+                    expectSpyCall(isQueryResultNotEmptySpy, 3, emptyQueryResult);
+                    expectToBe(component.isQueryResultNotEmpty(emptyQueryResult), false);
+                });
+
+                it('... queryResult.head.vars & queryResult.body.bindings are empty arrays', () => {
+                    expectSpyCall(isQueryResultNotEmptySpy, 2, expectedQueryResult);
+
+                    // Mock empty response
+                    const emptyQueryResult = { head: { vars: [] }, body: { bindings: [] } };
+                    component.queryResult$ = observableOf(emptyQueryResult);
+
+                    detectChangesOnPush(fixture);
+
+                    expectSpyCall(isQueryResultNotEmptySpy, 3, emptyQueryResult);
+                    expectToBe(component.isQueryResultNotEmpty(emptyQueryResult), false);
+                });
+            });
+
+            it('... should return true if queryResult.head.vars && queryResult.body.bindings is not empty (length > 0)', () => {
+                expectSpyCall(isQueryResultNotEmptySpy, 2, expectedQueryResult);
+
+                // Mock non-empty response
+                const nonEmptyQueryResult = {
+                    head: { vars: ['AnotherTestHeader'] },
+                    body: { bindings: [{ testKey: 'AnotherTestValue' }] },
+                };
+                component.queryResult$ = observableOf(nonEmptyQueryResult);
                 detectChangesOnPush(fixture);
 
-                // Panel body
-                const bodyDes = getAndExpectDebugElementByCss(
-                    compDe,
-                    'div#awg-graph-visualizer-select-results > div.accordion-body',
-                    1,
-                    1
-                );
-
-                getAndExpectDebugElementByDirective(bodyDes[0], TwelveToneSpinnerStubComponent, 1, 1);
-            });
-
-            it('... should contain panel body with SparqlNoResultsStubComponent (stubbed) if no results are available', () => {
-                // Mock empty response
-                component.queryResult$ = observableOf({ head: { vars: [] }, body: { bindings: [] } });
-                detectChangesOnPush(fixture);
-
-                // Panel body
-                const bodyDes = getAndExpectDebugElementByCss(
-                    compDe,
-                    'div#awg-graph-visualizer-select-results > div.accordion-body',
-                    1,
-                    1
-                );
-
-                // SparqlNoResultsStubComponent
-                getAndExpectDebugElementByDirective(bodyDes[0], SparqlNoResultsStubComponent, 1, 1);
-            });
-
-            it('... should contain panel body with SparqlTableComponent (stubbed) if results are available', () => {
-                // Panel body
-                const bodyDes = getAndExpectDebugElementByCss(
-                    compDe,
-                    'div#awg-graph-visualizer-select-results > div.accordion-body',
-                    1,
-                    1
-                );
-
-                // SparqlTable
-                getAndExpectDebugElementByDirective(bodyDes[0], SparqlTableStubComponent, 1, 1);
-            });
-
-            it('... should pass down `queryResult` and `queryTime` to sparqlTable component', () => {
-                const sparqlTableDes = getAndExpectDebugElementByDirective(compDe, SparqlTableStubComponent, 1, 1);
-                const sparqlTableCmp = sparqlTableDes[0].injector.get(
-                    SparqlTableStubComponent
-                ) as SparqlTableStubComponent;
-
-                expect(sparqlTableCmp.queryResult).toBeDefined();
-                expect(sparqlTableCmp.queryResult)
-                    .withContext(`should equal ${expectedQueryResult}`)
-                    .toEqual(expectedQueryResult);
-
-                expect(sparqlTableCmp.queryTime).toBeDefined();
-                expect(sparqlTableCmp.queryTime)
-                    .withContext(`should have data: ${expectedQueryTime}`)
-                    .toEqual(expectedQueryTime);
+                expectSpyCall(isQueryResultNotEmptySpy, 3, nonEmptyQueryResult);
+                expectToBe(component.isQueryResultNotEmpty(nonEmptyQueryResult), true);
             });
         });
 
@@ -390,193 +837,6 @@ describe('SelectResultsComponent (DONE)', () => {
 
                 expectSpyCall(tableClickSpy, 1, expectedUri);
                 expectSpyCall(emitClickedTableRequestSpy, 1, expectedUri);
-            });
-        });
-
-        describe('#isNotEmpty()', () => {
-            it('... should have a method `isNotEmpty`', () => {
-                expect(component.isNotEmpty).toBeDefined();
-            });
-
-            it('... should not do anything if no queryResult.head is provided', () => {
-                expectSpyCall(isNotEmptySpy, 1, expectedQueryResult);
-
-                // Mock empty response
-                const emptyQueryResult = { head: undefined, body: { bindings: [{ test: 'Test' }] } };
-                component.queryResult$ = observableOf(emptyQueryResult);
-                detectChangesOnPush(fixture);
-
-                expectSpyCall(isNotEmptySpy, 2, emptyQueryResult);
-                // SparqlNoResultsStubComponent
-                getAndExpectDebugElementByDirective(compDe, SparqlNoResultsStubComponent, 1, 1);
-            });
-
-            it('... should not do anything if no queryResult.body is provided', () => {
-                expectSpyCall(isNotEmptySpy, 1, expectedQueryResult);
-
-                // Mock empty response
-                const emptyQueryResult = { head: { vars: ['Test'] }, body: undefined };
-                component.queryResult$ = observableOf(emptyQueryResult);
-                detectChangesOnPush(fixture);
-
-                expectSpyCall(isNotEmptySpy, 2, emptyQueryResult);
-                // SparqlNoResultsStubComponent
-                getAndExpectDebugElementByDirective(compDe, SparqlNoResultsStubComponent, 1, 1);
-            });
-
-            it('... should return false if queryResult.head.vars length = 0', () => {
-                expectSpyCall(isNotEmptySpy, 1, expectedQueryResult);
-
-                // Mock empty response
-                const emptyQueryResult = { head: { vars: [] }, body: { bindings: [{ testKey: 'TestValue' }] } };
-                component.queryResult$ = observableOf(emptyQueryResult);
-                detectChangesOnPush(fixture);
-
-                expectSpyCall(isNotEmptySpy, 2, emptyQueryResult);
-                expect(component.isNotEmpty(emptyQueryResult)).toBeFalse();
-            });
-
-            it('... should return false if queryResult.body.bindings length = 0', () => {
-                expectSpyCall(isNotEmptySpy, 1, expectedQueryResult);
-
-                // Mock empty response
-                const emptyQueryResult = { head: { vars: ['TestHeader'] }, body: { bindings: [] } };
-                component.queryResult$ = observableOf(emptyQueryResult);
-                detectChangesOnPush(fixture);
-
-                expectSpyCall(isNotEmptySpy, 2, emptyQueryResult);
-                expect(component.isNotEmpty(emptyQueryResult)).toBeFalse();
-            });
-
-            it('... should return false if queryResult.head.vars & queryResult.body.bindings length = 0', () => {
-                expectSpyCall(isNotEmptySpy, 1, expectedQueryResult);
-
-                // Mock empty response
-                const emptyQueryResult = { head: { vars: [] }, body: { bindings: [] } };
-                component.queryResult$ = observableOf(emptyQueryResult);
-                detectChangesOnPush(fixture);
-
-                expectSpyCall(isNotEmptySpy, 2, emptyQueryResult);
-                expect(component.isNotEmpty(emptyQueryResult)).toBeFalse();
-            });
-
-            it('... should return true if queryResult.head.vars & queryResult.body.bindings length > 0', () => {
-                component.queryResult$ = expectedQueryResult$;
-                expectSpyCall(isNotEmptySpy, 1, expectedQueryResult);
-
-                expect(component.isNotEmpty(expectedQueryResult)).toBeTrue();
-                expectSpyCall(isNotEmptySpy, 2, expectedQueryResult);
-
-                // Mock another non-empty response
-                const queryResult = { head: { vars: ['TestHeader'] }, body: { bindings: [{ testKey: 'TestValue' }] } };
-                component.queryResult$ = observableOf(queryResult);
-                detectChangesOnPush(fixture);
-
-                expectSpyCall(isNotEmptySpy, 3, queryResult);
-                expect(component.isNotEmpty(queryResult)).toBeTrue();
-            });
-        });
-
-        describe('#preventPanelCollapseOnFullscreen()', () => {
-            it('... should have a method `preventPanelCollapseOnFullscreen`', () => {
-                expect(component.preventPanelCollapseOnFullscreen).toBeDefined();
-            });
-
-            it('... should trigger on event from ngb-accordion', () => {
-                const accordionDes = getAndExpectDebugElementByDirective(compDe, NgbAccordion, 1, 1);
-                const accordionCmp = accordionDes[0].injector.get(NgbAccordion) as NgbAccordion;
-
-                const panelChangeEvent: NgbPanelChangeEvent = {
-                    panelId: 'panelChangeId',
-                    nextState: true,
-                    preventDefault: () => {
-                        // Intentional empty test override
-                    },
-                };
-
-                // Emit change event from accordion
-                accordionCmp.panelChange.emit(panelChangeEvent);
-
-                expectSpyCall(preventPanelCollapseOnFullscreenSpy, 1, panelChangeEvent);
-            });
-
-            it('... should not do anything if no $event is provided', () => {
-                const accordionDes = getAndExpectDebugElementByDirective(compDe, NgbAccordion, 1, 1);
-                const accordionCmp = accordionDes[0].injector.get(NgbAccordion) as NgbAccordion;
-
-                // Emit undefined change event from accordion
-                accordionCmp.panelChange.emit(undefined);
-
-                expectSpyCall(preventPanelCollapseOnFullscreenSpy, 1, undefined);
-            });
-
-            it('... should trigger $event.preventDefault() if isFullscreen == true && $event.nextState == false', () => {
-                const accordionDes = getAndExpectDebugElementByDirective(compDe, NgbAccordion, 1, 1);
-                const accordionCmp = accordionDes[0].injector.get(NgbAccordion) as NgbAccordion;
-
-                const panelChangeEvent: NgbPanelChangeEvent = {
-                    panelId: 'panelChangeId',
-                    nextState: false,
-                    preventDefault: () => {
-                        // Intentional empty test override
-                    },
-                };
-                const preventDefaultSpy: Spy = spyOn(panelChangeEvent, 'preventDefault').and.callThrough();
-
-                // Set fullscreen mode
-                component.isFullscreen = true;
-
-                // Emit change event from accordion
-                accordionCmp.panelChange.emit(panelChangeEvent);
-
-                expectSpyCall(preventPanelCollapseOnFullscreenSpy, 1, panelChangeEvent);
-                expectSpyCall(preventDefaultSpy, 1, undefined);
-            });
-
-            it('... should not trigger $event.preventDefault() if $event.nextState == true', () => {
-                const accordionDes = getAndExpectDebugElementByDirective(compDe, NgbAccordion, 1, 1);
-                const accordionCmp = accordionDes[0].injector.get(NgbAccordion) as NgbAccordion;
-
-                const panelChangeEvent: NgbPanelChangeEvent = {
-                    panelId: 'panelChangeId',
-                    nextState: true,
-                    preventDefault: () => {
-                        // Intentional empty test override
-                    },
-                };
-                const preventDefaultSpy: Spy = spyOn(panelChangeEvent, 'preventDefault').and.callThrough();
-
-                // Set fullscreen mode
-                component.isFullscreen = true;
-
-                // Emit change event from accordion
-                accordionCmp.panelChange.emit(panelChangeEvent);
-
-                expectSpyCall(preventPanelCollapseOnFullscreenSpy, 1, panelChangeEvent);
-                expectSpyCall(preventDefaultSpy, 0, undefined);
-            });
-
-            it('... should not trigger $event.preventDefault() if isFullscreen == false', () => {
-                const accordionDes = getAndExpectDebugElementByDirective(compDe, NgbAccordion, 1, 1);
-                const accordionCmp = accordionDes[0].injector.get(NgbAccordion) as NgbAccordion;
-
-                const panelChangeEvent: NgbPanelChangeEvent = {
-                    panelId: 'panelChangeId',
-                    nextState: false,
-                    preventDefault: () => {
-                        // Intentional empty test override
-                    },
-                };
-                const preventDefaultSpy: Spy = spyOn(panelChangeEvent, 'preventDefault').and.callThrough();
-
-                // Unset fullscreen mode
-                component.isFullscreen = false;
-
-                // Emit change event from accordion
-                accordionCmp.panelChange.emit(panelChangeEvent);
-
-                expectSpyCall(preventPanelCollapseOnFullscreenSpy, 1, panelChangeEvent);
-                expectSpyCall(preventDefaultSpy, 0, undefined);
             });
         });
     });
