@@ -1,14 +1,13 @@
 import { Component, DebugElement, EventEmitter, Input, NgModule, Output, SimpleChange } from '@angular/core';
-import { ComponentFixture, fakeAsync, TestBed, tick, waitForAsync } from '@angular/core/testing';
+import { ComponentFixture, TestBed, fakeAsync, tick, waitForAsync } from '@angular/core/testing';
 
 import { sparql } from '@codemirror/legacy-modes/mode/sparql';
 import {
-    NgbAccordion,
+    NgbAccordionDirective,
     NgbAccordionModule,
     NgbConfig,
     NgbDropdown,
     NgbDropdownModule,
-    NgbPanelChangeEvent,
 } from '@ng-bootstrap/ng-bootstrap';
 import { EditorView } from 'codemirror';
 import Spy = jasmine.Spy;
@@ -17,6 +16,9 @@ import { click } from '@testing/click-helper';
 import { detectChangesOnPush } from '@testing/detect-changes-on-push-helper';
 import {
     expectSpyCall,
+    expectToBe,
+    expectToContain,
+    expectToEqual,
     getAndExpectDebugElementByCss,
     getAndExpectDebugElementByDirective,
 } from '@testing/expect-helper';
@@ -65,7 +67,8 @@ describe('SparqlEditorComponent (DONE)', () => {
     let onQueryListChangeSpy: Spy;
     let onViewChangeSpy: Spy;
     let performQuerySpy: Spy;
-    let preventPanelCollapseOnFullscreenSpy: Spy;
+    let isAccordionItemDisabledSpy: Spy;
+    let isAccordionItemCollapsedSpy: Spy;
     let resetQuerySpy: Spy;
     let setViewTypeSpy: Spy;
     let switchQueryTypeSpy: Spy;
@@ -85,7 +88,7 @@ describe('SparqlEditorComponent (DONE)', () => {
 
     beforeEach(waitForAsync(() => {
         TestBed.configureTestingModule({
-            imports: [NgbAnimationConfigModule, NgbAccordion, NgbDropdown],
+            imports: [NgbAnimationConfigModule, NgbAccordionDirective, NgbDropdown],
             declarations: [SparqlEditorComponent, CodeMirrorStubComponent, ViewHandleButtongGroupStubComponent],
         }).compileComponents();
     }));
@@ -96,6 +99,9 @@ describe('SparqlEditorComponent (DONE)', () => {
         compDe = fixture.debugElement;
 
         // Test data
+        expectedIsFullscreen = false;
+        expectedCmSparqlMode = sparql;
+
         expectedConstructQuery1 = new GraphSparqlQuery();
         expectedConstructQuery1.queryType = 'construct';
         expectedConstructQuery1.queryLabel = 'Test Query 3';
@@ -123,10 +129,6 @@ describe('SparqlEditorComponent (DONE)', () => {
             expectedSelectQuery2,
         ];
 
-        expectedIsFullscreen = false;
-
-        expectedCmSparqlMode = sparql;
-
         expectedViewHandles = [
             new ViewHandle('Graph view', ViewHandleTypes.GRAPH, component.faDiagramProject),
             new ViewHandle('Table view', ViewHandleTypes.TABLE, component.faTable),
@@ -140,7 +142,8 @@ describe('SparqlEditorComponent (DONE)', () => {
         onQueryListChangeSpy = spyOn(component, 'onQueryListChange').and.callThrough();
         onViewChangeSpy = spyOn(component, 'onViewChange').and.callThrough();
         performQuerySpy = spyOn(component, 'performQuery').and.callThrough();
-        preventPanelCollapseOnFullscreenSpy = spyOn(component, 'preventPanelCollapseOnFullscreen').and.callThrough();
+        isAccordionItemCollapsedSpy = spyOn(component, 'isAccordionItemCollapsed').and.callThrough();
+        isAccordionItemDisabledSpy = spyOn(component, 'isAccordionItemDisabled').and.callThrough();
         resetQuerySpy = spyOn(component, 'resetQuery').and.callThrough();
         setViewTypeSpy = spyOn(component, 'setViewType').and.callThrough();
         switchQueryTypeSpy = spyOn(component, 'switchQueryType').and.callThrough();
@@ -168,33 +171,21 @@ describe('SparqlEditorComponent (DONE)', () => {
         });
 
         it('... should have cmSparqlMode', () => {
-            expect(component.cmSparqlMode).toBeDefined();
-            expect(component.cmSparqlMode)
-                .withContext(`should equal ${expectedCmSparqlMode}`)
-                .toEqual(expectedCmSparqlMode);
+            expectToEqual(component.cmSparqlMode, expectedCmSparqlMode);
         });
 
         it('... should have selectedViewType', () => {
-            expect(component.selectedViewType).toBeDefined();
-            expect(component.selectedViewType)
-                .withContext(`should equal ${ViewHandleTypes.GRAPH}`)
-                .toEqual(ViewHandleTypes.GRAPH);
+            expectToEqual(component.selectedViewType, ViewHandleTypes.GRAPH);
         });
 
         it('... should have viewHandles', () => {
-            expect(component.viewHandles).toBeDefined();
-            expect(component.viewHandles)
-                .withContext(`should equal ${expectedViewHandles}`)
-                .toEqual(expectedViewHandles);
+            expectToEqual(component.viewHandles, expectedViewHandles);
         });
 
         describe('VIEW', () => {
-            it('... should contain one ngb-accordion without panel (div.accordion-item) yet', () => {
-                // Ngb-accordion debug element
-                const accordionDes = getAndExpectDebugElementByCss(compDe, 'ngb-accordion', 1, 1);
-
-                // Panel
-                getAndExpectDebugElementByCss(accordionDes[0], 'div.accordion-item', 0, 0, 'yet');
+            it('... should contain no ngbAccordion yet', () => {
+                // NgbAccordion debug element
+                getAndExpectDebugElementByCss(compDe, 'div.accordion', 0, 0);
             });
         });
     });
@@ -211,54 +202,72 @@ describe('SparqlEditorComponent (DONE)', () => {
         });
 
         it('... should have `queryList` input', () => {
-            expect(component.queryList).toBeDefined();
-            expect(component.queryList).withContext(`should equal ${expectedQueryList}`).toEqual(expectedQueryList);
+            expectToEqual(component.queryList, expectedQueryList);
         });
 
         it('... should have `query` input', () => {
-            expect(component.query).toBeDefined();
-            expect(component.query)
-                .withContext(`should equal ${expectedConstructQuery1}`)
-                .toEqual(expectedConstructQuery1);
+            expectToEqual(component.query, expectedConstructQuery1);
         });
 
         it('... should have `isFullScreen` input', () => {
-            expect(component.isFullscreen).toBeDefined();
-            expect(component.isFullscreen)
-                .withContext(`should equal ${expectedIsFullscreen}`)
-                .toBe(expectedIsFullscreen);
+            expectToBe(component.isFullscreen, expectedIsFullscreen);
         });
 
         describe('VIEW', () => {
+            it('... should contain one div.accordion', () => {
+                // NgbAccordion debug element
+                getAndExpectDebugElementByCss(compDe, 'div.accordion', 1, 1);
+            });
+
             describe('not in fullscreen mode', () => {
                 describe('with closed panel', () => {
-                    it('... should contain one ngb-accordion with panel (div.accordion-item) header and collapsed body', () => {
-                        // Ngb-accordion debug element
-                        const accordionDes = getAndExpectDebugElementByCss(compDe, 'ngb-accordion', 1, 1);
+                    it('... should contain one div.accordion with panel (div.accordion-item) header and collapsed body', () => {
+                        // NgbAccordion debug element
+                        const accordionDes = getAndExpectDebugElementByCss(compDe, 'div.accordion', 1, 1);
 
-                        // Panel (div.card)
-                        const panelDes = getAndExpectDebugElementByCss(accordionDes[0], 'div.accordion-item', 1, 1); // Panel (div.card)
-                        // Header
-                        getAndExpectDebugElementByCss(
-                            panelDes[0],
-                            'div#awg-graph-visualizer-sparql-query-header.accordion-header',
+                        // Panel (div.accordion-item)
+                        const panelDes = getAndExpectDebugElementByCss(
+                            accordionDes[0],
+                            'div#awg-graph-visualizer-sparql-query.accordion-item',
                             1,
                             1
-                        ); // Panel (div.card)
-                        // No body
-                        getAndExpectDebugElementByCss(
-                            panelDes[0],
-                            'div#awg-graph-visualizer-sparql-query > div.accordion-body',
-                            0,
-                            0
                         );
+                        // Header (div.accordion-header)
+                        const panelHeaderDes = getAndExpectDebugElementByCss(
+                            panelDes[0],
+                            'div#awg-graph-visualizer-sparql-query > div.accordion-header',
+                            1,
+                            1
+                        );
+                        const panelHeaderEl = panelHeaderDes[0].nativeElement;
+
+                        expectToContain(panelHeaderEl.classList, 'collapsed');
+
+                        // Body (div.accordion-collapse)
+                        const panelBodyDes = getAndExpectDebugElementByCss(
+                            panelDes[0],
+                            'div#awg-graph-visualizer-sparql-query > div.accordion-collapse',
+                            1,
+                            1
+                        );
+                        const panelBodyEl = panelBodyDes[0].nativeElement;
+
+                        expect(panelBodyEl.classList).not.toContain('show');
                     });
 
                     it('... should display panel header button', () => {
-                        // Panel header button
-                        const btnDes = getAndExpectDebugElementByCss(
+                        // Header debug elements
+                        const panelHeaderDes = getAndExpectDebugElementByCss(
                             compDe,
-                            'div#awg-graph-visualizer-sparql-query-header > div.accordion-button > button.btn-link',
+                            'div#awg-graph-visualizer-sparql-query > div.accordion-header',
+                            1,
+                            1
+                        );
+
+                        // Custom panel header button
+                        const btnDes = getAndExpectDebugElementByCss(
+                            panelHeaderDes[0],
+                            'button#awg-graph-visualizer-sparql-query-toggle',
                             1,
                             1
                         );
@@ -266,15 +275,14 @@ describe('SparqlEditorComponent (DONE)', () => {
                         const btnEl = btnDes[0].nativeElement;
 
                         // Check button content
-                        expect(btnEl.textContent).toBeTruthy();
-                        expect(btnEl.textContent).withContext('should be SPARQL').toContain('SPARQL');
+                        expectToBe(btnEl.textContent, 'SPARQL');
                     });
 
-                    it('... should toggle panel body by click on panel header button', async () => {
+                    it('... should toggle panel body on click', () => {
                         // Header debug elements
                         const panelHeaderDes = getAndExpectDebugElementByCss(
                             compDe,
-                            'div#awg-graph-visualizer-sparql-query-header.accordion-header',
+                            'div#awg-graph-visualizer-sparql-query > div.accordion-header',
                             1,
                             1
                         );
@@ -282,7 +290,7 @@ describe('SparqlEditorComponent (DONE)', () => {
                         // Button debug elements
                         const btnDes = getAndExpectDebugElementByCss(
                             panelHeaderDes[0],
-                            'div.accordion-button > button.btn-link',
+                            'button#awg-graph-visualizer-sparql-query-toggle',
                             1,
                             1
                         );
@@ -290,39 +298,45 @@ describe('SparqlEditorComponent (DONE)', () => {
                         const btnEl = btnDes[0].nativeElement;
 
                         // Panel body is closed
-                        getAndExpectDebugElementByCss(
+                        let panelBodyDes = getAndExpectDebugElementByCss(
                             compDe,
-                            'div#awg-graph-visualizer-sparql-query > div.accordion-body',
-                            0,
-                            0,
-                            'collapsed'
+                            'div#awg-graph-visualizer-sparql-query > div.accordion-collapse',
+                            1,
+                            1
                         );
+                        let panelBodyEl = panelBodyDes[0].nativeElement;
+
+                        expect(panelBodyEl.classList).not.toContain('show');
 
                         // Click header button
                         click(btnEl as HTMLElement);
-                        await detectChangesOnPush(fixture);
+                        detectChangesOnPush(fixture);
 
                         // Panel is open
-                        getAndExpectDebugElementByCss(
+                        panelBodyDes = getAndExpectDebugElementByCss(
                             compDe,
-                            'div#awg-graph-visualizer-sparql-query > div.accordion-body',
+                            'div#awg-graph-visualizer-sparql-query > div.accordion-collapse',
                             1,
-                            1,
-                            'open'
+                            1
                         );
+                        panelBodyEl = panelBodyDes[0].nativeElement;
+
+                        expectToContain(panelBodyEl.classList, 'show');
 
                         // Click header button
                         click(btnEl as HTMLElement);
-                        await detectChangesOnPush(fixture);
+                        detectChangesOnPush(fixture);
 
                         // Panel body is closed again
-                        getAndExpectDebugElementByCss(
+                        panelBodyDes = getAndExpectDebugElementByCss(
                             compDe,
-                            'div#awg-graph-visualizer-sparql-query > div.accordion-body',
-                            0,
-                            0,
-                            'collapsed'
+                            'div#awg-graph-visualizer-sparql-query > div.accordion-collapse',
+                            1,
+                            1
                         );
+                        panelBodyEl = panelBodyDes[0].nativeElement;
+
+                        expect(panelBodyEl.classList).not.toContain('show');
                     });
 
                     describe('View handle button group', () => {
@@ -330,7 +344,7 @@ describe('SparqlEditorComponent (DONE)', () => {
                             // Header debug elements
                             const panelHeaderDes = getAndExpectDebugElementByCss(
                                 compDe,
-                                'div#awg-graph-visualizer-sparql-query-header.accordion-header',
+                                'div#awg-graph-visualizer-sparql-query > div.accordion-header',
                                 1,
                                 1
                             );
@@ -347,7 +361,7 @@ describe('SparqlEditorComponent (DONE)', () => {
                             // Header debug elements
                             const panelHeaderDes = getAndExpectDebugElementByCss(
                                 compDe,
-                                'div#awg-graph-visualizer-sparql-query-header.accordion-header',
+                                'div#awg-graph-visualizer-sparql-query > div.accordion-header',
                                 1,
                                 1
                             );
@@ -362,21 +376,20 @@ describe('SparqlEditorComponent (DONE)', () => {
                                 ViewHandleButtongGroupStubComponent
                             ) as ViewHandleButtongGroupStubComponent;
 
-                            expect(viewHandleButtongGroupCmp.selectedViewType).toBeTruthy();
-                            expect(viewHandleButtongGroupCmp.selectedViewType).toBe(ViewHandleTypes.GRAPH);
+                            expectToBe(viewHandleButtongGroupCmp.selectedViewType, ViewHandleTypes.GRAPH);
                         });
                     });
 
                     describe('Example query button group', () => {
-                        it('... should contain an example query btn-group in panel header if isExampleQueriesEnabled = true', async () => {
+                        it('... should contain an example query btn-group in panel header if isExampleQueriesEnabled = true', () => {
                             isExampleQueriesEnabledSpy.and.returnValue(true);
 
-                            await detectChangesOnPush(fixture);
+                            detectChangesOnPush(fixture);
 
                             // Header debug elements
                             const panelHeaderDes = getAndExpectDebugElementByCss(
                                 compDe,
-                                'div#awg-graph-visualizer-sparql-query-header.accordion-header',
+                                'div#awg-graph-visualizer-sparql-query > div.accordion-header',
                                 1,
                                 1
                             );
@@ -384,21 +397,21 @@ describe('SparqlEditorComponent (DONE)', () => {
                             // Panel header div.awg-example-query-btn-group
                             getAndExpectDebugElementByCss(
                                 panelHeaderDes[0],
-                                'div.accordion-button > div.awg-example-query-btn-group',
+                                'div.accordion-header > div.awg-example-query-btn-group',
                                 1,
                                 1
                             );
                         });
 
-                        it('... should not contain an example query btn-group in panel header if isExampleQueriesEnabled = false', async () => {
+                        it('... should not contain an example query btn-group in panel header if isExampleQueriesEnabled = false', () => {
                             isExampleQueriesEnabledSpy.and.returnValue(false);
 
-                            await detectChangesOnPush(fixture);
+                            detectChangesOnPush(fixture);
 
                             // Header debug elements
                             const panelHeaderDes = getAndExpectDebugElementByCss(
                                 compDe,
-                                'div#awg-graph-visualizer-sparql-query-header.accordion-header',
+                                'div#awg-graph-visualizer-sparql-query > div.accordion-header',
                                 1,
                                 1
                             );
@@ -406,7 +419,7 @@ describe('SparqlEditorComponent (DONE)', () => {
                             // Panel header div.awg-example-query-btn-group
                             getAndExpectDebugElementByCss(
                                 panelHeaderDes[0],
-                                'div.accordion-button > div.awg-example-query-btn-group',
+                                'div.accordion-header > div.awg-example-query-btn-group',
                                 0,
                                 0
                             );
@@ -416,38 +429,36 @@ describe('SparqlEditorComponent (DONE)', () => {
                             // Header debug elements
                             const panelHeaderDes = getAndExpectDebugElementByCss(
                                 compDe,
-                                'div#awg-graph-visualizer-sparql-query-header.accordion-header',
+                                'div#awg-graph-visualizer-sparql-query > div.accordion-header',
                                 1,
                                 1
                             );
 
                             const btnDes = getAndExpectDebugElementByCss(
                                 panelHeaderDes[0],
-                                'div.accordion-button > div.awg-example-query-btn-group > button.btn',
+                                'div.accordion-header > div.awg-example-query-btn-group > button.btn',
                                 1,
                                 1
                             );
                             const btnEl = btnDes[0].nativeElement;
 
                             expect(btnEl.disabled).toBeTruthy();
-                            expect(btnEl.getAttribute('aria-disabled')).toBe('true');
-
-                            expect(btnEl.textContent).toBeTruthy();
-                            expect(btnEl.textContent).toContain('Beispielabfragen');
+                            expectToBe(btnEl.getAttribute('aria-disabled'), 'true');
+                            expectToBe(btnEl.textContent.trim(), 'Beispielabfragen');
                         });
 
                         it('... should contain another btn-group dropdown in example query btn-group', () => {
                             // Header debug elements
                             const panelHeaderDes = getAndExpectDebugElementByCss(
                                 compDe,
-                                'div#awg-graph-visualizer-sparql-query-header.accordion-header',
+                                'div#awg-graph-visualizer-sparql-query > div.accordion-header',
                                 1,
                                 1
                             );
 
                             const outerButtonGroupDes = getAndExpectDebugElementByCss(
                                 panelHeaderDes[0],
-                                'div.accordion-button > div.awg-example-query-btn-group',
+                                'div.accordion-header > div.awg-example-query-btn-group',
                                 1,
                                 1
                             );
@@ -459,14 +470,14 @@ describe('SparqlEditorComponent (DONE)', () => {
                             // Header debug elements
                             const panelHeaderDes = getAndExpectDebugElementByCss(
                                 compDe,
-                                'div#awg-graph-visualizer-sparql-query-header.accordion-header',
+                                'div#awg-graph-visualizer-sparql-query > div.accordion-header',
                                 1,
                                 1
                             );
 
                             const dropdownButtonGroupDes = getAndExpectDebugElementByCss(
                                 panelHeaderDes[0],
-                                'div.accordion-button > div.awg-example-query-btn-group > div.btn-group.dropdown',
+                                'div.accordion-header > div.awg-example-query-btn-group > div.btn-group.dropdown',
                                 1,
                                 1
                             );
@@ -484,14 +495,14 @@ describe('SparqlEditorComponent (DONE)', () => {
                             // Header debug elements
                             const panelHeaderDes = getAndExpectDebugElementByCss(
                                 compDe,
-                                'div#awg-graph-visualizer-sparql-query-header.accordion-header',
+                                'div#awg-graph-visualizer-sparql-query > div.accordion-header',
                                 1,
                                 1
                             );
 
                             const dropdownButtonGroupDes = getAndExpectDebugElementByCss(
                                 panelHeaderDes[0],
-                                'div.accordion-button > div.awg-example-query-btn-group > div.btn-group.dropdown',
+                                'div.accordion-header > div.awg-example-query-btn-group > div.btn-group.dropdown',
                                 1,
                                 1
                             );
@@ -514,7 +525,7 @@ describe('SparqlEditorComponent (DONE)', () => {
                         it('... should display label on dropdown items in example query btn-group', () => {
                             const dropdownButtonGroupDes = getAndExpectDebugElementByCss(
                                 compDe,
-                                'div.accordion-button > div.awg-example-query-btn-group > div.btn-group.dropdown',
+                                'div.accordion-header > div.awg-example-query-btn-group > div.btn-group.dropdown',
                                 1,
                                 1
                             );
@@ -526,25 +537,16 @@ describe('SparqlEditorComponent (DONE)', () => {
                                 expectedQueryList.length
                             );
 
-                            expect(itemDes.length)
-                                .withContext(`should be ${expectedQueryList.length}`)
-                                .toBe(expectedQueryList.length);
+                            expectToBe(itemDes.length, expectedQueryList.length);
 
-                            const itemEl0 = itemDes[0].nativeElement;
-                            const itemEl1 = itemDes[1].nativeElement;
+                            itemDes.forEach((itemDe: DebugElement, index: number) => {
+                                const itemEl = itemDe.nativeElement;
 
-                            expect(itemEl0.textContent).toBeTruthy();
-                            expect(itemEl0.textContent)
-                                .withContext(`should contain ${expectedQueryList[0].queryLabel}`)
-                                .toContain(expectedQueryList[0].queryLabel);
-
-                            expect(itemEl1.textContent).toBeTruthy();
-                            expect(itemEl1.textContent)
-                                .withContext(`should contain ${expectedQueryList[1].queryLabel}`)
-                                .toContain(expectedQueryList[1].queryLabel);
+                                expectToBe(itemEl.textContent, expectedQueryList[index].queryLabel);
+                            });
                         });
 
-                        it('... should disable current query in dropdown items', async () => {
+                        it('... should disable current query in dropdown items', () => {
                             const itemDes = getAndExpectDebugElementByCss(
                                 compDe,
                                 'div.dropdown-menu > a.dropdown-item',
@@ -552,9 +554,7 @@ describe('SparqlEditorComponent (DONE)', () => {
                                 expectedQueryList.length
                             );
 
-                            expect(itemDes.length)
-                                .withContext(`should be ${expectedQueryList.length}`)
-                                .toBe(expectedQueryList.length);
+                            expectToBe(itemDes.length, expectedQueryList.length);
 
                             const itemEl0 = itemDes[0].nativeElement;
                             const itemEl1 = itemDes[1].nativeElement;
@@ -563,13 +563,13 @@ describe('SparqlEditorComponent (DONE)', () => {
                             expect(itemEl1).not.toHaveClass('disabled');
 
                             component.query = expectedConstructQuery2;
-                            await detectChangesOnPush(fixture);
+                            detectChangesOnPush(fixture);
 
                             expect(itemEl0).not.toHaveClass('disabled');
                             expect(itemEl1).toHaveClass('disabled');
                         });
 
-                        it('... should trigger `onQueryListChange()` by click on dropdown item', async () => {
+                        it('... should trigger `onQueryListChange()` by click on dropdown item', () => {
                             const itemDes = getAndExpectDebugElementByCss(
                                 compDe,
                                 'div.dropdown-menu > a.dropdown-item',
@@ -577,9 +577,7 @@ describe('SparqlEditorComponent (DONE)', () => {
                                 expectedQueryList.length
                             );
 
-                            expect(itemDes.length)
-                                .withContext(`should be ${expectedQueryList.length}`)
-                                .toBe(expectedQueryList.length);
+                            expectToBe(itemDes.length, expectedQueryList.length);
 
                             const itemEl0 = itemDes[0].nativeElement;
                             const itemEl1 = itemDes[1].nativeElement;
@@ -588,19 +586,19 @@ describe('SparqlEditorComponent (DONE)', () => {
 
                             // Click on second item (first disabled)
                             click(itemEl1 as HTMLElement);
-                            await detectChangesOnPush(fixture);
+                            detectChangesOnPush(fixture);
 
                             // Spy call with second query
                             expectSpyCall(onQueryListChangeSpy, 1, expectedConstructQuery2);
 
                             component.query = expectedConstructQuery2;
-                            await detectChangesOnPush(fixture);
+                            detectChangesOnPush(fixture);
 
                             expect(itemEl0).not.toHaveClass('disabled');
 
                             // Click on first item (second disabled)
                             click(itemEl0 as HTMLElement);
-                            await detectChangesOnPush(fixture);
+                            detectChangesOnPush(fixture);
 
                             // Spy call with first query
                             expectSpyCall(onQueryListChangeSpy, 2, expectedConstructQuery1);
@@ -611,11 +609,11 @@ describe('SparqlEditorComponent (DONE)', () => {
                 describe('with open panel', () => {
                     let bodyDes: DebugElement[];
 
-                    beforeEach(async () => {
+                    beforeEach(() => {
                         // Open panel by click on header button
                         const btnDes = getAndExpectDebugElementByCss(
                             compDe,
-                            'div#awg-graph-visualizer-sparql-query-header.accordion-header > div.accordion-button > button.btn-link',
+                            'button#awg-graph-visualizer-sparql-query-toggle',
                             1,
                             1
                         );
@@ -623,15 +621,82 @@ describe('SparqlEditorComponent (DONE)', () => {
 
                         // Click header button
                         click(btnEl as HTMLElement);
-                        await detectChangesOnPush(fixture);
+                        detectChangesOnPush(fixture);
 
-                        // Panel body
-                        bodyDes = getAndExpectDebugElementByCss(
+                        // Panel body is open
+                        const collapseDes = getAndExpectDebugElementByCss(
                             compDe,
-                            'div#awg-graph-visualizer-sparql-query > div.accordion-body',
+                            'div#awg-graph-visualizer-sparql-query > div.accordion-collapse',
                             1,
                             1
                         );
+                        const collapseEl = collapseDes[0].nativeElement;
+
+                        expectToContain(collapseEl.classList, 'show');
+
+                        // Panel body
+                        bodyDes = getAndExpectDebugElementByCss(collapseDes[0], 'div.accordion-body', 1, 1);
+                    });
+
+                    it('... should toggle panel body on click', () => {
+                        // Header debug elements
+                        const panelHeaderDes = getAndExpectDebugElementByCss(
+                            compDe,
+                            'div#awg-graph-visualizer-sparql-query > div.accordion-header',
+                            1,
+                            1
+                        );
+
+                        // Button debug elements
+                        const btnDes = getAndExpectDebugElementByCss(
+                            panelHeaderDes[0],
+                            'button#awg-graph-visualizer-sparql-query-toggle',
+                            1,
+                            1
+                        );
+                        // Button native elements to click on
+                        const btnEl = btnDes[0].nativeElement;
+
+                        // Panel body is open
+                        let panelBodyDes = getAndExpectDebugElementByCss(
+                            compDe,
+                            'div#awg-graph-visualizer-sparql-query > div.accordion-collapse',
+                            1,
+                            1
+                        );
+                        let panelBodyEl = panelBodyDes[0].nativeElement;
+
+                        expectToContain(panelBodyEl.classList, 'show');
+
+                        // Click header button
+                        click(btnEl as HTMLElement);
+                        detectChangesOnPush(fixture);
+
+                        // Panel is closed
+                        panelBodyDes = getAndExpectDebugElementByCss(
+                            compDe,
+                            'div#awg-graph-visualizer-sparql-query > div.accordion-collapse',
+                            1,
+                            1
+                        );
+                        panelBodyEl = panelBodyDes[0].nativeElement;
+
+                        expect(panelBodyEl.classList).not.toContain('show');
+
+                        // Click header button
+                        click(btnEl as HTMLElement);
+                        detectChangesOnPush(fixture);
+
+                        // Panel body is open again
+                        panelBodyDes = getAndExpectDebugElementByCss(
+                            compDe,
+                            'div#awg-graph-visualizer-sparql-query > div.accordion-collapse',
+                            1,
+                            1
+                        );
+                        panelBodyEl = panelBodyDes[0].nativeElement;
+
+                        expectToContain(panelBodyEl.classList, 'show');
                     });
 
                     it('... should contain CodeMirrorComponent (stubbed) in panel body', () => {
@@ -652,17 +717,12 @@ describe('SparqlEditorComponent (DONE)', () => {
                         const btnEl1 = btnDes[1].nativeElement;
                         const btnEl2 = btnDes[2].nativeElement;
 
-                        expect(btnEl0.textContent).toBeTruthy();
-                        expect(btnEl0.textContent).withContext(`should contain 'Query'`).toContain('Query');
-
-                        expect(btnEl1.textContent).toBeTruthy();
-                        expect(btnEl1.textContent).withContext(`should contain 'Reset'`).toContain('Reset');
-
-                        expect(btnEl2.textContent).toBeTruthy();
-                        expect(btnEl2.textContent).withContext(`should contain 'Clear'`).toContain('Clear');
+                        expectToBe(btnEl0.textContent, 'Query');
+                        expectToBe(btnEl1.textContent, 'Reset');
+                        expectToBe(btnEl2.textContent, 'Clear');
                     });
 
-                    it('... should trigger `performQuery()` by click on Query button', async () => {
+                    it('... should trigger `performQuery()` by click on Query button', () => {
                         const btnDes = getAndExpectDebugElementByCss(
                             bodyDes[0],
                             'div.awg-graph-visualizer-sparql-query-handle-buttons > button.btn',
@@ -671,18 +731,17 @@ describe('SparqlEditorComponent (DONE)', () => {
                         );
                         const btnEl0 = btnDes[0].nativeElement;
 
-                        expect(btnEl0.textContent).toBeTruthy();
-                        expect(btnEl0.textContent).withContext(`should contain 'Query'`).toContain('Query');
+                        expectToBe(btnEl0.textContent, 'Query');
 
                         // Click query button
                         click(btnEl0 as HTMLElement);
-                        await detectChangesOnPush(fixture);
+                        detectChangesOnPush(fixture);
 
                         expectSpyCall(performQuerySpy, 1);
                         expectSpyCall(resetQuerySpy, 0);
                     });
 
-                    it('... should trigger `resetQuery()` by click on Reset button', async () => {
+                    it('... should trigger `resetQuery()` by click on Reset button', () => {
                         const btnDes = getAndExpectDebugElementByCss(
                             bodyDes[0],
                             'div.awg-graph-visualizer-sparql-query-handle-buttons > button.btn',
@@ -691,18 +750,17 @@ describe('SparqlEditorComponent (DONE)', () => {
                         );
                         const btnEl1 = btnDes[1].nativeElement;
 
-                        expect(btnEl1.textContent).toBeTruthy();
-                        expect(btnEl1.textContent).withContext(`should contain 'Reset'`).toContain('Reset');
+                        expectToBe(btnEl1.textContent, 'Reset');
 
                         // Click reset button
                         click(btnEl1 as HTMLElement);
-                        await detectChangesOnPush(fixture);
+                        detectChangesOnPush(fixture);
 
                         expectSpyCall(performQuerySpy, 0);
                         expectSpyCall(resetQuerySpy, 1);
                     });
 
-                    it('... should trigger `onEditorInputChange()` with empty string by click on Clear button', async () => {
+                    it('... should trigger `onEditorInputChange()` with empty string by click on Clear button', () => {
                         const btnDes = getAndExpectDebugElementByCss(
                             bodyDes[0],
                             'div.awg-graph-visualizer-sparql-query-handle-buttons > button.btn',
@@ -711,12 +769,11 @@ describe('SparqlEditorComponent (DONE)', () => {
                         );
                         const btnEl2 = btnDes[2].nativeElement;
 
-                        expect(btnEl2.textContent).toBeTruthy();
-                        expect(btnEl2.textContent).withContext(`should contain 'Clear'`).toContain('Clear');
+                        expectToBe(btnEl2.textContent, 'Clear');
 
                         // Click clear button
                         click(btnEl2 as HTMLElement);
-                        await detectChangesOnPush(fixture);
+                        detectChangesOnPush(fixture);
 
                         expectSpyCall(onEditorInputChangeSpy, 1, '');
                     });
@@ -730,7 +787,7 @@ describe('SparqlEditorComponent (DONE)', () => {
                     // Open panel by click on header button
                     const btnDes = getAndExpectDebugElementByCss(
                         compDe,
-                        'div#awg-graph-visualizer-sparql-query-header.accordion-header > div.accordion-button > button.btn-link',
+                        'button#awg-graph-visualizer-sparql-query-toggle',
                         1,
                         1
                     );
@@ -743,32 +800,39 @@ describe('SparqlEditorComponent (DONE)', () => {
                     // Panel body
                     bodyDes = getAndExpectDebugElementByCss(
                         compDe,
-                        'div#awg-graph-visualizer-sparql-query > div.accordion-body',
+                        'div#awg-graph-visualizer-sparql-query-collapse > div.accordion-body',
                         1,
                         1
                     );
 
                     // Set fullscreen mode
                     component.isFullscreen = true;
+                    detectChangesOnPush(fixture);
                 });
 
-                it('... should contain one ngb-accordion with panel (div.accordion-item) header and open body', () => {
-                    // Ngb-accordion debug element
-                    const accordionDes = getAndExpectDebugElementByCss(compDe, 'ngb-accordion', 1, 1);
+                it('... should contain one div.accordion with panel (div.accordion-item) header and open body', () => {
+                    // NgbAccordion debug element
+                    const accordionDes = getAndExpectDebugElementByCss(compDe, 'div.accordion', 1, 1);
 
-                    // Panel (div.card)
-                    const panelDes = getAndExpectDebugElementByCss(accordionDes[0], 'div.accordion-item', 1, 1); // Panel (div.card)
-                    // Header
-                    getAndExpectDebugElementByCss(
-                        panelDes[0],
-                        'div#awg-graph-visualizer-sparql-query-header.accordion-header',
+                    // Panel (div.accordion-item)
+                    const panelDes = getAndExpectDebugElementByCss(
+                        accordionDes[0],
+                        'div#awg-graph-visualizer-sparql-query.accordion-item',
                         1,
                         1
-                    ); // Panel (div.card)
-                    // Body open
+                    );
+                    // Header (div.accordion-header)
                     getAndExpectDebugElementByCss(
                         panelDes[0],
-                        'div#awg-graph-visualizer-sparql-query > div.accordion-body',
+                        'div#awg-graph-visualizer-sparql-query > div.accordion-header',
+                        1,
+                        1
+                    );
+
+                    // Body open (div.accordion-body)
+                    getAndExpectDebugElementByCss(
+                        panelDes[0],
+                        'div#awg-graph-visualizer-sparql-query-collapse > div.accordion-body',
                         1,
                         1
                     );
@@ -778,32 +842,29 @@ describe('SparqlEditorComponent (DONE)', () => {
                     // Header debug elements
                     const panelHeaderDes = getAndExpectDebugElementByCss(
                         compDe,
-                        'div#awg-graph-visualizer-sparql-query-header.accordion-header',
+                        'div#awg-graph-visualizer-sparql-query > div.accordion-header',
                         1,
                         1
                     );
 
-                    // Panel header button
+                    // Custom panel header button
                     const btnDes = getAndExpectDebugElementByCss(
                         panelHeaderDes[0],
-                        'div.accordion-button > button.btn-link',
+                        'button#awg-graph-visualizer-sparql-query-toggle',
                         1,
                         1
                     );
-
-                    // Button native element
                     const btnEl = btnDes[0].nativeElement;
 
                     // Check button content
-                    expect(btnEl.textContent).toBeTruthy();
-                    expect(btnEl.textContent).withContext(`should be 'SPARQL'`).toContain('SPARQL');
+                    expectToBe(btnEl.textContent, 'SPARQL');
                 });
 
-                it('... should not toggle panel body by click on panel header button', async () => {
+                it('... should not toggle panel body on click', () => {
                     // Header debug elements
                     const panelHeaderDes = getAndExpectDebugElementByCss(
                         compDe,
-                        'div#awg-graph-visualizer-sparql-query-header.accordion-header',
+                        'div#awg-graph-visualizer-sparql-query > div.accordion-header',
                         1,
                         1
                     );
@@ -811,17 +872,17 @@ describe('SparqlEditorComponent (DONE)', () => {
                     // Button debug element
                     const btnDes = getAndExpectDebugElementByCss(
                         panelHeaderDes[0],
-                        'div.accordion-button > button.btn-link',
+                        'div.accordion-button > button#awg-graph-visualizer-sparql-query-toggle',
                         1,
                         1
                     );
                     // Button native element to click on
                     const btnEl = btnDes[0].nativeElement;
 
-                    // Panel body is open
+                    // Panel body does not close
                     getAndExpectDebugElementByCss(
                         compDe,
-                        'div#awg-graph-visualizer-sparql-query > div.accordion-body',
+                        'div#awg-graph-visualizer-sparql-query-collapse > div.accordion-body',
                         1,
                         1,
                         'open'
@@ -829,12 +890,12 @@ describe('SparqlEditorComponent (DONE)', () => {
 
                     // Click header button
                     click(btnEl as HTMLElement);
-                    await detectChangesOnPush(fixture);
+                    detectChangesOnPush(fixture);
 
                     // Panel is open again
                     getAndExpectDebugElementByCss(
                         compDe,
-                        'div#awg-graph-visualizer-sparql-query > div.accordion-body',
+                        'div#awg-graph-visualizer-sparql-query-collapse > div.accordion-body',
                         1,
                         1,
                         'open'
@@ -842,11 +903,11 @@ describe('SparqlEditorComponent (DONE)', () => {
                 });
 
                 describe('View handle button group', () => {
-                    it('... should contain ViewHandleButtongGroupComponent (stubbed) in panel header', () => {
+                    it('... should contain no ViewHandleButtongGroupComponent (stubbed) in panel header', () => {
                         // Header debug elements
                         const panelHeaderDes = getAndExpectDebugElementByCss(
                             compDe,
-                            'div#awg-graph-visualizer-sparql-query-header.accordion-header',
+                            'div#awg-graph-visualizer-sparql-query > div.accordion-header',
                             1,
                             1
                         );
@@ -854,45 +915,22 @@ describe('SparqlEditorComponent (DONE)', () => {
                         getAndExpectDebugElementByDirective(
                             panelHeaderDes[0],
                             ViewHandleButtongGroupStubComponent,
-                            1,
-                            1
+                            0,
+                            0
                         );
-                    });
-
-                    it('... should have selectedViewType===graph (according to querytype)', () => {
-                        // Header debug elements
-                        const panelHeaderDes = getAndExpectDebugElementByCss(
-                            compDe,
-                            'div#awg-graph-visualizer-sparql-query-header.accordion-header',
-                            1,
-                            1
-                        );
-                        // ViewHandleButtongGroupComponent debug elements
-                        const viewHandleButtongGroupDes = getAndExpectDebugElementByDirective(
-                            panelHeaderDes[0],
-                            ViewHandleButtongGroupStubComponent,
-                            1,
-                            1
-                        );
-                        const viewHandleButtongGroupCmp = viewHandleButtongGroupDes[0].injector.get(
-                            ViewHandleButtongGroupStubComponent
-                        ) as ViewHandleButtongGroupStubComponent;
-
-                        expect(viewHandleButtongGroupCmp.selectedViewType).toBeTruthy();
-                        expect(viewHandleButtongGroupCmp.selectedViewType).toBe(ViewHandleTypes.GRAPH);
                     });
                 });
 
                 describe('Example query button group', () => {
-                    it('... should contain an example query btn-group in panel header if isExampleQueriesEnabled = true', async () => {
+                    it('... should contain an example query btn-group in panel header if isExampleQueriesEnabled = true', () => {
                         isExampleQueriesEnabledSpy.and.returnValue(true);
 
-                        await detectChangesOnPush(fixture);
+                        detectChangesOnPush(fixture);
 
                         // Header debug elements
                         const panelHeaderDes = getAndExpectDebugElementByCss(
                             compDe,
-                            'div#awg-graph-visualizer-sparql-query-header.accordion-header',
+                            'div#awg-graph-visualizer-sparql-query > div.accordion-header',
                             1,
                             1
                         );
@@ -906,15 +944,15 @@ describe('SparqlEditorComponent (DONE)', () => {
                         );
                     });
 
-                    it('... should not contain an example query btn-group in panel header if isExampleQueriesEnabled = false', async () => {
+                    it('... should not contain an example query btn-group in panel header if isExampleQueriesEnabled = false', () => {
                         isExampleQueriesEnabledSpy.and.returnValue(false);
 
-                        await detectChangesOnPush(fixture);
+                        detectChangesOnPush(fixture);
 
                         // Header debug elements
                         const panelHeaderDes = getAndExpectDebugElementByCss(
                             compDe,
-                            'div#awg-graph-visualizer-sparql-query-header.accordion-header',
+                            'div#awg-graph-visualizer-sparql-query > div.accordion-header',
                             1,
                             1
                         );
@@ -932,7 +970,7 @@ describe('SparqlEditorComponent (DONE)', () => {
                         // Header debug elements
                         const panelHeaderDes = getAndExpectDebugElementByCss(
                             compDe,
-                            'div#awg-graph-visualizer-sparql-query-header.accordion-header',
+                            'div#awg-graph-visualizer-sparql-query > div.accordion-header',
                             1,
                             1
                         );
@@ -946,17 +984,15 @@ describe('SparqlEditorComponent (DONE)', () => {
                         const btnEl = btnDes[0].nativeElement;
 
                         expect(btnEl.disabled).toBeTruthy();
-                        expect(btnEl.getAttribute('aria-disabled')).toBe('true');
-
-                        expect(btnEl.textContent).toBeTruthy();
-                        expect(btnEl.textContent).toContain('Beispielabfragen');
+                        expectToBe(btnEl.getAttribute('aria-disabled'), 'true');
+                        expectToBe(btnEl.textContent.trim(), 'Beispielabfragen');
                     });
 
                     it('... should contain another btn-group dropdown in example query btn-group', () => {
                         // Header debug elements
                         const panelHeaderDes = getAndExpectDebugElementByCss(
                             compDe,
-                            'div#awg-graph-visualizer-sparql-query-header.accordion-header',
+                            'div#awg-graph-visualizer-sparql-query > div.accordion-header',
                             1,
                             1
                         );
@@ -975,7 +1011,7 @@ describe('SparqlEditorComponent (DONE)', () => {
                         // Header debug elements
                         const panelHeaderDes = getAndExpectDebugElementByCss(
                             compDe,
-                            'div#awg-graph-visualizer-sparql-query-header.accordion-header',
+                            'div#awg-graph-visualizer-sparql-query > div.accordion-header',
                             1,
                             1
                         );
@@ -994,7 +1030,7 @@ describe('SparqlEditorComponent (DONE)', () => {
                         // Header debug elements
                         const panelHeaderDes = getAndExpectDebugElementByCss(
                             compDe,
-                            'div#awg-graph-visualizer-sparql-query-header.accordion-header',
+                            'div#awg-graph-visualizer-sparql-query > div.accordion-header',
                             1,
                             1
                         );
@@ -1036,25 +1072,16 @@ describe('SparqlEditorComponent (DONE)', () => {
                             expectedQueryList.length
                         );
 
-                        expect(itemDes.length)
-                            .withContext(`should be ${expectedQueryList.length}`)
-                            .toBe(expectedQueryList.length);
+                        expectToBe(itemDes.length, expectedQueryList.length);
 
-                        const itemEl0 = itemDes[0].nativeElement;
-                        const itemEl1 = itemDes[1].nativeElement;
+                        itemDes.forEach((itemDe: DebugElement, index: number) => {
+                            const itemEl = itemDe.nativeElement;
 
-                        expect(itemEl0.textContent).toBeTruthy();
-                        expect(itemEl0.textContent)
-                            .withContext(`should contain ${expectedQueryList[0].queryLabel}`)
-                            .toContain(expectedQueryList[0].queryLabel);
-
-                        expect(itemEl1.textContent).toBeTruthy();
-                        expect(itemEl1.textContent)
-                            .withContext(`should contain ${expectedQueryList[1].queryLabel}`)
-                            .toContain(expectedQueryList[1].queryLabel);
+                            expectToBe(itemEl.textContent, expectedQueryList[index].queryLabel);
+                        });
                     });
 
-                    it('... should disable current query in dropdown items in example query btn-group', async () => {
+                    it('... should disable current query in dropdown items in example query btn-group', () => {
                         const itemDes = getAndExpectDebugElementByCss(
                             compDe,
                             'div.dropdown-menu > a.dropdown-item',
@@ -1062,9 +1089,7 @@ describe('SparqlEditorComponent (DONE)', () => {
                             expectedQueryList.length
                         );
 
-                        expect(itemDes.length)
-                            .withContext(`should be ${expectedQueryList.length}`)
-                            .toBe(expectedQueryList.length);
+                        expectToBe(itemDes.length, expectedQueryList.length);
 
                         const itemEl0 = itemDes[0].nativeElement;
                         const itemEl1 = itemDes[1].nativeElement;
@@ -1073,13 +1098,13 @@ describe('SparqlEditorComponent (DONE)', () => {
                         expect(itemEl1).not.toHaveClass('disabled');
 
                         component.query = expectedConstructQuery2;
-                        await detectChangesOnPush(fixture);
+                        detectChangesOnPush(fixture);
 
                         expect(itemEl0).not.toHaveClass('disabled');
                         expect(itemEl1).toHaveClass('disabled');
                     });
 
-                    it('... should trigger `onQueryListChange()` by click on dropdown item', async () => {
+                    it('... should trigger `onQueryListChange()` by click on dropdown item', () => {
                         const itemDes = getAndExpectDebugElementByCss(
                             compDe,
                             'div.dropdown-menu > a.dropdown-item',
@@ -1087,9 +1112,7 @@ describe('SparqlEditorComponent (DONE)', () => {
                             expectedQueryList.length
                         );
 
-                        expect(itemDes.length)
-                            .withContext(`should be ${expectedQueryList.length}`)
-                            .toBe(expectedQueryList.length);
+                        expectToBe(itemDes.length, expectedQueryList.length);
 
                         const itemEl0 = itemDes[0].nativeElement;
                         const itemEl1 = itemDes[1].nativeElement;
@@ -1098,19 +1121,19 @@ describe('SparqlEditorComponent (DONE)', () => {
 
                         // Click on second item (first disabled)
                         click(itemEl1 as HTMLElement);
-                        await detectChangesOnPush(fixture);
+                        detectChangesOnPush(fixture);
 
                         // Spy call with second query
                         expectSpyCall(onQueryListChangeSpy, 1, expectedConstructQuery2);
 
                         component.query = expectedConstructQuery2;
-                        await detectChangesOnPush(fixture);
+                        detectChangesOnPush(fixture);
 
                         expect(itemEl0).not.toHaveClass('disabled');
 
                         // Click on first item (second disabled)
                         click(itemEl0 as HTMLElement);
-                        await detectChangesOnPush(fixture);
+                        detectChangesOnPush(fixture);
 
                         // Spy call with first query
                         expectSpyCall(onQueryListChangeSpy, 2, expectedConstructQuery1);
@@ -1123,83 +1146,55 @@ describe('SparqlEditorComponent (DONE)', () => {
                 });
 
                 it('... should contain div with 3 buttons (Query, Reset, Clear) in panel body', () => {
-                    const divDes = getAndExpectDebugElementByCss(
-                        bodyDes[0],
-                        'div.awg-graph-visualizer-sparql-query-handle-buttons',
-                        1,
-                        1
-                    );
+                    const divDes = getAndExpectDebugElementByCss(bodyDes[0], 'div', 1, 1);
 
                     const btnDes = getAndExpectDebugElementByCss(divDes[0], 'button.btn', 3, 3);
                     const btnEl0 = btnDes[0].nativeElement;
                     const btnEl1 = btnDes[1].nativeElement;
                     const btnEl2 = btnDes[2].nativeElement;
 
-                    expect(btnEl0.textContent).toBeTruthy();
-                    expect(btnEl0.textContent).withContext(`should contain 'Query'`).toContain('Query');
-
-                    expect(btnEl1.textContent).toBeTruthy();
-                    expect(btnEl1.textContent).withContext(`should contain 'Reset'`).toContain('Reset');
-
-                    expect(btnEl2.textContent).toBeTruthy();
-                    expect(btnEl2.textContent).withContext(`should contain 'Clear'`).toContain('Clear');
+                    expectToBe(btnEl0.textContent, 'Query');
+                    expectToBe(btnEl1.textContent, 'Reset');
+                    expectToBe(btnEl2.textContent, 'Clear');
                 });
 
-                it('... should trigger `performQuery()` by click on Query button', async () => {
-                    const btnDes = getAndExpectDebugElementByCss(
-                        bodyDes[0],
-                        'div.awg-graph-visualizer-sparql-query-handle-buttons > button.btn',
-                        3,
-                        3
-                    );
+                it('... should trigger `performQuery()` by click on Query button', () => {
+                    const btnDes = getAndExpectDebugElementByCss(bodyDes[0], 'div > button.btn', 3, 3);
                     const btnEl0 = btnDes[0].nativeElement;
 
-                    expect(btnEl0.textContent).toBeTruthy();
-                    expect(btnEl0.textContent).withContext(`should contain 'Query'`).toContain('Query');
+                    expectToBe(btnEl0.textContent, 'Query');
 
                     // Click query button
                     click(btnEl0 as HTMLElement);
-                    await detectChangesOnPush(fixture);
+                    detectChangesOnPush(fixture);
 
                     expectSpyCall(performQuerySpy, 1);
                     expectSpyCall(resetQuerySpy, 0);
                 });
 
-                it('... should trigger `resetQuery()` by click on Reset button', async () => {
-                    const btnDes = getAndExpectDebugElementByCss(
-                        bodyDes[0],
-                        'div.awg-graph-visualizer-sparql-query-handle-buttons > button.btn',
-                        3,
-                        3
-                    );
+                it('... should trigger `resetTriples()` by click on Reset button', () => {
+                    const btnDes = getAndExpectDebugElementByCss(bodyDes[0], 'div > button.btn', 3, 3);
                     const btnEl1 = btnDes[1].nativeElement;
 
-                    expect(btnEl1.textContent).toBeTruthy();
-                    expect(btnEl1.textContent).withContext(`should contain 'Reset'`).toContain('Reset');
+                    expectToBe(btnEl1.textContent, 'Reset');
 
                     // Click reset button
                     click(btnEl1 as HTMLElement);
-                    await detectChangesOnPush(fixture);
+                    detectChangesOnPush(fixture);
 
                     expectSpyCall(performQuerySpy, 0);
                     expectSpyCall(resetQuerySpy, 1);
                 });
 
-                it('... should trigger `onEditorInputChange()` with empty string by click on Clear button', async () => {
-                    const btnDes = getAndExpectDebugElementByCss(
-                        bodyDes[0],
-                        'div.awg-graph-visualizer-sparql-query-handle-buttons > button.btn',
-                        3,
-                        3
-                    );
+                it('... should trigger `onEditorInputChange()` with empty string by click on Clear button', () => {
+                    const btnDes = getAndExpectDebugElementByCss(bodyDes[0], 'div > button.btn', 3, 3);
                     const btnEl2 = btnDes[2].nativeElement;
 
-                    expect(btnEl2.textContent).toBeTruthy();
-                    expect(btnEl2.textContent).withContext(`should contain 'Clear'`).toContain('Clear');
+                    expectToBe(btnEl2.textContent, 'Clear');
 
                     // Click clear button
                     click(btnEl2 as HTMLElement);
-                    await detectChangesOnPush(fixture);
+                    detectChangesOnPush(fixture);
 
                     expectSpyCall(onEditorInputChangeSpy, 1, '');
                 });
@@ -1276,7 +1271,7 @@ describe('SparqlEditorComponent (DONE)', () => {
                 // Open panel by click on header button
                 const btnDes = getAndExpectDebugElementByCss(
                     compDe,
-                    'div#awg-graph-visualizer-sparql-query-header.accordion-header > div.accordion-button > button.btn-link',
+                    'div#awg-graph-visualizer-sparql-query > div.accordion-header > button.btn-link',
                     1,
                     1
                 );
@@ -1301,7 +1296,7 @@ describe('SparqlEditorComponent (DONE)', () => {
                 expectSpyCall(onEditorInputChangeSpy, 1, changedQueryString);
             });
 
-            it('... should trigger with empty string from click on Clear button', async () => {
+            it('... should trigger with empty string from click on Clear button', () => {
                 const btnDes = getAndExpectDebugElementByCss(
                     compDe,
                     'div.awg-graph-visualizer-sparql-query-handle-buttons > button.btn',
@@ -1310,17 +1305,16 @@ describe('SparqlEditorComponent (DONE)', () => {
                 );
                 const btnEl2 = btnDes[2].nativeElement;
 
-                expect(btnEl2.textContent).toBeTruthy();
-                expect(btnEl2.textContent).withContext(`should contain 'Clear'`).toContain('Clear');
+                expectToBe(btnEl2.textContent, 'Clear');
 
                 // Click clear button
                 click(btnEl2 as HTMLElement);
-                await detectChangesOnPush(fixture);
+                detectChangesOnPush(fixture);
 
                 expectSpyCall(onEditorInputChangeSpy, 1, '');
             });
 
-            it('... should emit updateQueryStringRequest on click', async () => {
+            it('... should emit updateQueryStringRequest on click', () => {
                 const btnDes = getAndExpectDebugElementByCss(
                     compDe,
                     'div.awg-graph-visualizer-sparql-query-handle-buttons > button.btn',
@@ -1329,12 +1323,11 @@ describe('SparqlEditorComponent (DONE)', () => {
                 );
                 const btnEl2 = btnDes[2].nativeElement;
 
-                expect(btnEl2.textContent).toBeTruthy();
-                expect(btnEl2.textContent).withContext(`should contain 'Clear'`).toContain('Clear');
+                expectToBe(btnEl2.textContent, 'Clear');
 
                 // Click clear button
                 click(btnEl2 as HTMLElement);
-                await detectChangesOnPush(fixture);
+                detectChangesOnPush(fixture);
 
                 expectSpyCall(onEditorInputChangeSpy, 1, '');
                 expectSpyCall(emitUpdateQueryStringRequestSpy, 1);
@@ -1476,7 +1469,7 @@ describe('SparqlEditorComponent (DONE)', () => {
                 // Header debug elements
                 const panelHeaderDes = getAndExpectDebugElementByCss(
                     compDe,
-                    'div#awg-graph-visualizer-sparql-query-header.accordion-header',
+                    'div#awg-graph-visualizer-sparql-query > div.accordion-header',
                     1,
                     1
                 );
@@ -1511,7 +1504,7 @@ describe('SparqlEditorComponent (DONE)', () => {
                 // Open panel by click on header button
                 const btnDes = getAndExpectDebugElementByCss(
                     compDe,
-                    'div#awg-graph-visualizer-sparql-query-header.accordion-header > div.accordion-button > button.btn-link',
+                    'div#awg-graph-visualizer-sparql-query > div.accordion-header > button.btn-link',
                     1,
                     1
                 );
@@ -1526,7 +1519,7 @@ describe('SparqlEditorComponent (DONE)', () => {
                 expect(component.performQuery).toBeDefined();
             });
 
-            it('... should trigger from click on Query button', async () => {
+            it('... should trigger from click on Query button', () => {
                 const btnDes = getAndExpectDebugElementByCss(
                     compDe,
                     'div.awg-graph-visualizer-sparql-query-handle-buttons > button.btn',
@@ -1535,18 +1528,17 @@ describe('SparqlEditorComponent (DONE)', () => {
                 );
                 const btnEl0 = btnDes[0].nativeElement;
 
-                expect(btnEl0.textContent).toBeTruthy();
-                expect(btnEl0.textContent).withContext(`should contain 'Query'`).toContain('Query');
+                expectToBe(btnEl0.textContent, 'Query');
 
                 // Click query button
                 click(btnEl0 as HTMLElement);
-                await detectChangesOnPush(fixture);
+                detectChangesOnPush(fixture);
 
                 expectSpyCall(performQuerySpy, 1);
             });
 
-            describe('... should emit on click:', async () => {
-                it('`performQueryRequest` if querystring is given', async () => {
+            describe('... should emit on click:', () => {
+                it('`performQueryRequest` if querystring is given', () => {
                     const btnDes = getAndExpectDebugElementByCss(
                         compDe,
                         'div.awg-graph-visualizer-sparql-query-handle-buttons > button.btn',
@@ -1555,12 +1547,11 @@ describe('SparqlEditorComponent (DONE)', () => {
                     );
                     const btnEl0 = btnDes[0].nativeElement;
 
-                    expect(btnEl0.textContent).toBeTruthy();
-                    expect(btnEl0.textContent).withContext(`should contain 'Query'`).toContain('Query');
+                    expectToBe(btnEl0.textContent, 'Query');
 
                     // Click query button
                     click(btnEl0 as HTMLElement);
-                    await detectChangesOnPush(fixture);
+                    detectChangesOnPush(fixture);
 
                     expectSpyCall(performQuerySpy, 1);
                     expectSpyCall(emitPerformQueryRequestSpy, 1);
@@ -1571,7 +1562,7 @@ describe('SparqlEditorComponent (DONE)', () => {
                     const expectedErrorMessage = new ToastMessage('Empty query', 'Please enter a SPARQL query.');
 
                     component.query.queryString = '';
-                    await detectChangesOnPush(fixture);
+                    detectChangesOnPush(fixture);
 
                     const btnDes = getAndExpectDebugElementByCss(
                         compDe,
@@ -1581,12 +1572,11 @@ describe('SparqlEditorComponent (DONE)', () => {
                     );
                     const btnEl0 = btnDes[0].nativeElement;
 
-                    expect(btnEl0.textContent).toBeTruthy();
-                    expect(btnEl0.textContent).withContext(`should contain 'Query'`).toContain('Query');
+                    expectToBe(btnEl0.textContent, 'Query');
 
                     // Click query button
                     click(btnEl0 as HTMLElement);
-                    await detectChangesOnPush(fixture);
+                    detectChangesOnPush(fixture);
 
                     expectSpyCall(performQuerySpy, 1);
                     expectSpyCall(emitPerformQueryRequestSpy, 0);
@@ -1600,7 +1590,7 @@ describe('SparqlEditorComponent (DONE)', () => {
                 // Open panel by click on header button
                 const btnDes = getAndExpectDebugElementByCss(
                     compDe,
-                    'div#awg-graph-visualizer-sparql-query-header.accordion-header > div.accordion-button > button.btn-link',
+                    'div#awg-graph-visualizer-sparql-query > div.accordion-header > button.btn-link',
                     1,
                     1
                 );
@@ -1615,7 +1605,7 @@ describe('SparqlEditorComponent (DONE)', () => {
                 expect(component.resetQuery).toBeDefined();
             });
 
-            it('... should trigger from click on Reset button', async () => {
+            it('... should trigger on click on Reset button', () => {
                 const btnDes = getAndExpectDebugElementByCss(
                     compDe,
                     'div.awg-graph-visualizer-sparql-query-handle-buttons > button.btn',
@@ -1624,26 +1614,25 @@ describe('SparqlEditorComponent (DONE)', () => {
                 );
                 const btnEl1 = btnDes[1].nativeElement;
 
-                expect(btnEl1.textContent).toBeTruthy();
-                expect(btnEl1.textContent).withContext(`should contain 'Reset'`).toContain('Reset');
+                expectToBe(btnEl1.textContent, 'Reset');
 
                 // Click query button
                 click(btnEl1 as HTMLElement);
-                await detectChangesOnPush(fixture);
+                detectChangesOnPush(fixture);
 
                 expectSpyCall(resetQuerySpy, 1);
             });
 
-            it('... should not emit anything if no query is provided', async () => {
+            it('... should not emit anything if no query is provided', () => {
                 // Query is undefined
                 component.resetQuery(undefined);
-                await detectChangesOnPush(fixture);
+                detectChangesOnPush(fixture);
 
                 expectSpyCall(resetQuerySpy, 1, undefined);
                 expectSpyCall(emitResestQueryRequestSpy, 0);
             });
 
-            it('... should emit request on click', async () => {
+            it('... should emit request on click', () => {
                 const btnDes = getAndExpectDebugElementByCss(
                     compDe,
                     'div.awg-graph-visualizer-sparql-query-handle-buttons > button.btn',
@@ -1652,12 +1641,11 @@ describe('SparqlEditorComponent (DONE)', () => {
                 );
                 const btnEl1 = btnDes[1].nativeElement;
 
-                expect(btnEl1.textContent).toBeTruthy();
-                expect(btnEl1.textContent).withContext(`should contain 'Reset'`).toContain('Reset');
+                expectToBe(btnEl1.textContent, 'Reset');
 
                 // Click reset button
                 click(btnEl1 as HTMLElement);
-                await detectChangesOnPush(fixture);
+                detectChangesOnPush(fixture);
 
                 expectSpyCall(resetQuerySpy, 1);
                 expectSpyCall(emitResestQueryRequestSpy, 1);
@@ -1731,14 +1719,8 @@ describe('SparqlEditorComponent (DONE)', () => {
 
                 component.switchQueryType(ViewHandleTypes.TABLE);
 
-                expect(component.query.queryType).toBeTruthy();
-                expect(component.query.queryType)
-                    .withContext(`should be ${expectedSelectQuery1.queryType}`)
-                    .toBe(expectedSelectQuery1.queryType);
-                expect(component.query.queryString).toBeTruthy();
-                expect(component.query.queryString)
-                    .withContext(`should be ${expectedSelectQuery1.queryString}`)
-                    .toBe(expectedSelectQuery1.queryString);
+                expectToBe(component.query.queryType, expectedSelectQuery1.queryType);
+                expectToBe(component.query.queryString, expectedSelectQuery1.queryString);
             });
 
             it('... should switch querytype and string to `construct` if requested view is `graph`', () => {
@@ -1751,14 +1733,8 @@ describe('SparqlEditorComponent (DONE)', () => {
                 // Switch back to GRAPH view
                 component.switchQueryType(ViewHandleTypes.GRAPH);
 
-                expect(component.query.queryType).toBeTruthy();
-                expect(component.query.queryType)
-                    .withContext(`should be ${expectedConstructQuery1.queryType}`)
-                    .toBe(expectedConstructQuery1.queryType);
-                expect(component.query.queryString).toBeTruthy();
-                expect(component.query.queryString)
-                    .withContext(`should be ${expectedConstructQuery1.queryString}`)
-                    .toBe(expectedConstructQuery1.queryString);
+                expectToBe(component.query.queryType, expectedConstructQuery1.queryType);
+                expectToBe(component.query.queryString, expectedConstructQuery1.queryString);
             });
 
             it('... should do nothing if requested view is `grid`', () => {
@@ -1767,14 +1743,8 @@ describe('SparqlEditorComponent (DONE)', () => {
 
                 component.switchQueryType(ViewHandleTypes.GRID);
 
-                expect(component.query.queryType).toBeTruthy();
-                expect(component.query.queryType)
-                    .withContext(`should be ${expectedConstructQuery1.queryType}`)
-                    .toBe(expectedConstructQuery1.queryType);
-                expect(component.query.queryString).toBeTruthy();
-                expect(component.query.queryString)
-                    .withContext(`should be ${expectedConstructQuery1.queryString}`)
-                    .toBe(expectedConstructQuery1.queryString);
+                expectToBe(component.query.queryType, expectedConstructQuery1.queryType);
+                expectToBe(component.query.queryString, expectedConstructQuery1.queryString);
             });
 
             it('... should throw error if requested view is not `table`, `graph` or `grid`', () => {
@@ -1787,124 +1757,45 @@ describe('SparqlEditorComponent (DONE)', () => {
             });
         });
 
-        describe('#togglePanel()', () => {
-            it('... should have a method `togglePanel()`', () => {
-                expect(component.togglePanel).toBeDefined();
+        describe('#isAccordionItemCollapsed()', () => {
+            it('... should have a method `isAccordionItemCollapsed`', () => {
+                expect(component.isAccordionItemCollapsed).toBeDefined();
             });
 
-            it('... should return empty string if isFullscreen = false', () => {
-                expect(component.togglePanel()).not.toBeTruthy();
+            it('... should be triggered from ngbAccordionItem', () => {
+                expectSpyCall(isAccordionItemCollapsedSpy, 1);
             });
 
-            it('... should return panel id if isFullscreen = true', () => {
-                const expectedId = 'awg-graph-visualizer-sparql-query';
+            it('... should return true if isFullscreen is false', () => {
+                expectToBe(component.isAccordionItemCollapsed(), true);
+            });
 
+            it('... should return false if isFullscreen is true', () => {
                 // Set fullscreen flag to true
                 component.isFullscreen = true;
-                expect(component.togglePanel()).withContext(`should be ${expectedId}`).toBe(expectedId);
+
+                expectToBe(component.isAccordionItemCollapsed(), false);
             });
         });
 
-        describe('#preventPanelCollapseOnFullscreen()', () => {
-            it('... should have a method `preventPanelCollapseOnFullscreen`', () => {
-                expect(component.preventPanelCollapseOnFullscreen).toBeDefined();
+        describe('#isAccordionItemDisabled()', () => {
+            it('... should have a method `isAccordionItemDisabled`', () => {
+                expect(component.isAccordionItemDisabled).toBeDefined();
             });
 
-            it('... should trigger on event from ngb-accordion', () => {
-                const accordionDes = getAndExpectDebugElementByDirective(compDe, NgbAccordion, 1, 1);
-                const accordionCmp = accordionDes[0].injector.get(NgbAccordion) as NgbAccordion;
-
-                const panelChangeEvent: NgbPanelChangeEvent = {
-                    panelId: 'panelChangeId',
-                    nextState: true,
-                    preventDefault: () => {
-                        // Intentional empty test override
-                    },
-                };
-
-                // Emit change event from accordion
-                accordionCmp.panelChange.emit(panelChangeEvent);
-
-                expectSpyCall(preventPanelCollapseOnFullscreenSpy, 1, panelChangeEvent);
+            it('... should be triggered from ngbAccordionItem', () => {
+                expectSpyCall(isAccordionItemDisabledSpy, 1);
             });
 
-            it('... should not do anything if no $event is provided', () => {
-                const accordionDes = getAndExpectDebugElementByDirective(compDe, NgbAccordion, 1, 1);
-                const accordionCmp = accordionDes[0].injector.get(NgbAccordion) as NgbAccordion;
-
-                // Emit undefined change event from accordion
-                accordionCmp.panelChange.emit(undefined);
-
-                expectSpyCall(preventPanelCollapseOnFullscreenSpy, 1, undefined);
+            it('... should return false if isFullscreen is false', () => {
+                expectToBe(component.isAccordionItemDisabled(), false);
             });
 
-            it('... should trigger $event.preventDefault() if isFullscreen == true && $event.nextState == false', () => {
-                const accordionDes = getAndExpectDebugElementByDirective(compDe, NgbAccordion, 1, 1);
-                const accordionCmp = accordionDes[0].injector.get(NgbAccordion) as NgbAccordion;
-
-                const panelChangeEvent: NgbPanelChangeEvent = {
-                    panelId: 'panelChangeId',
-                    nextState: false,
-                    preventDefault: () => {
-                        // Intentional empty test override
-                    },
-                };
-                const preventDefaultSpy: Spy = spyOn(panelChangeEvent, 'preventDefault').and.callThrough();
-
-                // Set fullscreen mode
+            it('... should return true if isFullscreen is true', () => {
+                // Set fullscreen flag to true
                 component.isFullscreen = true;
 
-                // Emit change event from accordion
-                accordionCmp.panelChange.emit(panelChangeEvent);
-
-                expectSpyCall(preventPanelCollapseOnFullscreenSpy, 1, panelChangeEvent);
-                expectSpyCall(preventDefaultSpy, 1, undefined);
-            });
-
-            it('... should not trigger $event.preventDefault() if $event.nextState == true', () => {
-                const accordionDes = getAndExpectDebugElementByDirective(compDe, NgbAccordion, 1, 1);
-                const accordionCmp = accordionDes[0].injector.get(NgbAccordion) as NgbAccordion;
-
-                const panelChangeEvent: NgbPanelChangeEvent = {
-                    panelId: 'panelChangeId',
-                    nextState: true,
-                    preventDefault: () => {
-                        // Intentional empty test override
-                    },
-                };
-                const preventDefaultSpy: Spy = spyOn(panelChangeEvent, 'preventDefault').and.callThrough();
-
-                // Set fullscreen mode
-                component.isFullscreen = true;
-
-                // Emit change event from accordion
-                accordionCmp.panelChange.emit(panelChangeEvent);
-
-                expectSpyCall(preventPanelCollapseOnFullscreenSpy, 1, panelChangeEvent);
-                expectSpyCall(preventDefaultSpy, 0, undefined);
-            });
-
-            it('... should not trigger $event.preventDefault() if isFullscreen == false', () => {
-                const accordionDes = getAndExpectDebugElementByDirective(compDe, NgbAccordion, 1, 1);
-                const accordionCmp = accordionDes[0].injector.get(NgbAccordion) as NgbAccordion;
-
-                const panelChangeEvent: NgbPanelChangeEvent = {
-                    panelId: 'panelChangeId',
-                    nextState: false,
-                    preventDefault: () => {
-                        // Intentional empty test override
-                    },
-                };
-                const preventDefaultSpy: Spy = spyOn(panelChangeEvent, 'preventDefault').and.callThrough();
-
-                // Unset fullscreen mode
-                component.isFullscreen = false;
-
-                // Emit change event from accordion
-                accordionCmp.panelChange.emit(panelChangeEvent);
-
-                expectSpyCall(preventPanelCollapseOnFullscreenSpy, 1, panelChangeEvent);
-                expectSpyCall(preventDefaultSpy, 0, undefined);
+                expectToBe(component.isAccordionItemDisabled(), true);
             });
         });
     });
