@@ -79,7 +79,10 @@ export class HomeViewComponent implements OnInit {
      * @param {CoreService} coreService Instance of the CoreService.
      * @param {Router} router Instance of the Angular router.
      */
-    constructor(private coreService: CoreService, private router: Router) {}
+    constructor(
+        private coreService: CoreService,
+        private router: Router
+    ) {}
 
     /**
      * Getter variable: editionRouteConstants.
