@@ -1,33 +1,49 @@
 import { DebugElement } from '@angular/core';
 
-import { getAndExpectDebugElementByCss } from '@testing/expect-helper';
-
 /**
- * Test helper function: expectClosedPanelBody.
+ * Test helper function: expectPanel.
  *
- * It checks if a given ngb-accordion panel is closed.
+ * It checks if a given ngbAccordion item (panel) is collapsed or open.
  *
- * @param {DebugElement} de The DebugElement.
- * @param {string} id The id of the panel.
+ * @param {DebugElement} headerDe The accordion header DebugElement.
  * @param {string} msg The msg to be displayed in Expectation.
+ * @param {boolean} collapsed The boolean value if the panel is collapsed.
  *
- * @returns {void} Sends the given values to getAndExpect helper.
+ * @returns {DebugElement} The debug element of the panel.
  */
-export function expectClosedPanelBody(de: DebugElement, id: string, msg: string) {
-    getAndExpectDebugElementByCss(de, `div.accordion-item > div#${id} > div.accordion-body`, 0, 0, msg);
+function expectPanel(headerDe: DebugElement, msg: string, collapsed?: boolean): void {
+    const headerEl = headerDe.nativeElement;
+    const isCollapsed = headerEl.classList.contains('collapsed');
+
+    expect(isCollapsed)
+        .withContext(`Header should be ${collapsed ? 'collapsed' : 'not collapsed'}`)
+        .toBe(collapsed);
 }
 
 /**
- * Test helper function: expectClosedPanelBody.
+ * Test helper function: expectCollapsedPanel.
  *
- * It checks if a given ngb-accordion panel is open.
+ * It checks if a given ngbAccordion item (panel) is collapsed.
  *
- * @param {DebugElement} de The DebugElement.
- * @param {string} id The id of the panel.
+ * @param {DebugElement} headerDe The accordion header DebugElement.
  * @param {string} msg The msg to be displayed in Expectation.
  *
  * @returns {void} Sends the given values to getAndExpect helper.
  */
-export function expectOpenPanelBody(de: DebugElement, id: string, msg: string) {
-    getAndExpectDebugElementByCss(de, `div.accordion-item > div#${id} > div.accordion-body`, 1, 1, msg);
+export function expectCollapsedPanel(headerDe: DebugElement, msg: string) {
+    expectPanel(headerDe, msg, true);
+}
+
+/**
+ * Test helper function: expectOpenPanel.
+ *
+ * It checks if a given ngbAccordion item (panel) is open.
+ *
+ * @param {DebugElement} headerDe The accordion header DebugElement.
+ * @param {string} msg The msg to be displayed in Expectation.
+ *
+ * @returns {void} Sends the given values to getAndExpect helper.
+ */
+export function expectOpenPanel(headerDe: DebugElement, msg: string) {
+    expectPanel(headerDe, msg, false);
 }
